refactor(tauri): split commands.rs into command-family submodules (TD-014)
Hub commands.rs re-exports all #[tauri::command] fns + helpers; logic moved to
commands/{admin,demo,input,movement,reactions,saves,screenshot,setup,snapshot}.rs.
Pre-verify snapshot from interrupted run; finish pass brings to green.

Part of #1200 (TD-014).
