+++ parish/crates/parish-tauri/src/commands.rs
@@ -2,3898 +2,59 @@
 //!
 //! Each public function here is registered with `tauri::generate_handler!` and
 //! becomes callable from the Svelte frontend via `invoke("command_name", args)`.
-
-use std::sync::Arc;
-
-use parish_core::config::InferenceCategory;
-use parish_core::debug_snapshot::{self, AuthDebug, DebugEvent, DebugSnapshot, InferenceDebug};
-// AnyClient, InferenceQueue, spawn_inference_worker formerly imported here —
-// now handled by parish_core::game_loop::rebuild_inference_worker (#696).
-use parish_core::input::{InputResult, classify_input, parse_intent};
-use parish_core::ipc::{compute_name_hints, text_log, text_log_for_stream_turn, text_log_typed};
-use parish_core::npc::reactions;
-use parish_core::world::LocationId;
-// DEFAULT_START_LOCATION — no longer used directly; handled by load_fresh_world_and_npcs (#696).
-use tauri::Emitter;
-
-use crate::events::{
-    EVENT_STREAM_END, EVENT_STREAM_TOKEN, EVENT_STREAM_TURN_END, EVENT_TEXT_LOG,
-    EVENT_TRAVEL_START, EVENT_WORLD_UPDATE, StreamEndPayload, StreamTokenPayload,
-    StreamTurnEndPayload, TextLogPayload,
-};
-use crate::{AppState, MapData, MapLocation, NpcInfo, SaveState, ThemePalette, WorldSnapshot};
-
-/// Returns a formatted game-time string (`HH:MM YYYY-MM-DD`) snapshotted
-/// from the shared world clock. Used for debug event timestamps so the
-/// Events tab no longer renders blank times.
-async fn debug_event_timestamp(state: &Arc<AppState>) -> String {
-    let world = state.world.lock().await;
-    world.clock.now().format("%H:%M %Y-%m-%d").to_string()
-}
-
-// ── Helper: build a WorldSnapshot from locked world state ────────────────────
-
-/// Builds a [`WorldSnapshot`] from a locked world state reference.
-///
-/// Used both by the `get_world_snapshot` command and by the background
-/// idle-tick task in `lib.rs`. Includes name pronunciation hints when
-/// NPC manager and pronunciation data are provided.
-pub fn get_world_snapshot_inner(
-    world: &parish_core::world::WorldState,
-    npc_manager: Option<&parish_core::npc::manager::NpcManager>,
-    pronunciations: &[parish_core::game_mod::PronunciationEntry],
-) -> WorldSnapshot {
-    let mut snapshot = snapshot_from_world(world);
-    if let Some(npc_mgr) = npc_manager {
-        snapshot.name_hints = compute_name_hints(world, npc_mgr, pronunciations);
-    }
-    snapshot
-}
-
-/// Converts a core [`parish_core::ipc::WorldSnapshot`] into the Tauri-specific
-/// [`WorldSnapshot`] (which includes additional fields like `name_hints`).
-fn snapshot_from_world(world: &parish_core::world::WorldState) -> WorldSnapshot {
-    let core = parish_core::ipc::snapshot_from_world(world);
-    WorldSnapshot {
-        location_name: core.location_name,
-        location_description: core.location_description,
-        time_label: core.time_label,
-        hour: core.hour,
-        minute: core.minute,
-        weather: core.weather,
-        season: core.season,
-        festival: core.festival,
-        paused: core.paused,
-        inference_paused: core.inference_paused,
-        game_epoch_ms: core.game_epoch_ms,
-        speed_factor: core.speed_factor,
-        name_hints: vec![],
-        day_of_week: core.day_of_week,
-        turn_in_flight: core.turn_in_flight,
-    }
-}
-
-// compute_name_hints is now shared via parish_core::ipc::compute_name_hints
-
-// ── Commands ─────────────────────────────────────────────────────────────────
-
-/// Returns a snapshot of the current world state (location, time, weather, season).
-#[tauri::command]
-pub async fn get_world_snapshot(
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<WorldSnapshot, String> {
-    let world = state.world.lock().await;
-    let npc_manager = state.npc_manager.lock().await;
-    let snapshot = get_world_snapshot_inner(&world, Some(&npc_manager), &state.pronunciations);
-    Ok(snapshot)
-}
-
-/// Returns the map data: visited locations with coordinates, edges, and player position.
-///
-/// Includes visited locations (fully enriched) and the frontier — unvisited
-/// locations adjacent to any visited location — so the player can see where
-/// to explore next. Frontier locations are marked with `visited: false`.
-#[tauri::command]
-pub async fn get_map(state: tauri::State<'_, Arc<AppState>>) -> Result<MapData, String> {
-    let world = state.world.lock().await;
-    let config = state.config.lock().await;
-    let transport = state.transport.default_mode();
-    let core_map =
-        parish_core::ipc::build_map_data(&world, transport, config.reveal_unexplored_locations);
-
-    let player_loc = world.player_location;
-    let (player_lat, player_lon) = world
-        .graph
-        .get(player_loc)
-        .map(|data| (data.lat, data.lon))
-        .unwrap_or((0.0, 0.0));
-
-    Ok(MapData {
-        locations: core_map
-            .locations
-            .into_iter()
-            .map(|l| MapLocation {
-                id: l.id,
-                name: l.name,
-                lat: l.lat,
-                lon: l.lon,
-                adjacent: l.adjacent,
-                hops: l.hops,
-                indoor: l.indoor,
-                travel_minutes: l.travel_minutes,
-                visited: l.visited,
-            })
-            .collect(),
-        edges: core_map.edges,
-        player_location: core_map.player_location,
-        player_lat,
-        player_lon,
-        edge_traversals: core_map.edge_traversals,
-        transport_label: core_map.transport_label,
-        transport_id: core_map.transport_id,
-    })
-}
-
-/// Returns the list of NPCs currently at the player's location.
-#[tauri::command]
-pub async fn get_npcs_here(state: tauri::State<'_, Arc<AppState>>) -> Result<Vec<NpcInfo>, String> {
-    let world = state.world.lock().await;
-    let npc_manager = state.npc_manager.lock().await;
-    Ok(parish_core::ipc::build_npcs_here(&world, &npc_manager))
-}
-
-/// Returns the current palette as CSS hex colours.
-///
-/// Resolution order:
-/// 1. Mod-provided time-of-day keyframes → interpolated palette for the
-///    current game hour.
-/// 2. Mod-provided static `[theme.palette]` (no keyframes) → returned as-is.
-/// 3. No mod loaded → `neutral_grey_palette()` so the prompt overlay renders.
-#[tauri::command]
-pub async fn get_theme(state: tauri::State<'_, Arc<AppState>>) -> Result<ThemePalette, String> {
-    use chrono::Timelike;
-    use parish_core::config::PaletteConfig;
-    use parish_palette::{compute_palette_with_keyframes, neutral_grey_palette};
-    let raw = if !state.theme_keyframes.is_empty() {
-        let world = state.world.lock().await;
-        let now = world.clock.now();
-        compute_palette_with_keyframes(
-            now.hour(),
-            now.minute(),
-            &state.theme_keyframes,
-            &PaletteConfig::default(),
-        )
-    } else if let Some(p) = state.static_raw_palette {
-        p
-    } else {
-        neutral_grey_palette()
-    };
-    Ok(ThemePalette::from(raw))
-}
-
-/// Returns a debug snapshot of all game state for the debug panel.
-///
-/// Aggregates clock, world graph, NPC state, events, and inference config
-/// into a single serializable [`DebugSnapshot`].
-#[tauri::command]
-pub async fn get_debug_snapshot(
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<DebugSnapshot, String> {
-    Ok(build_app_debug_snapshot(&state).await)
-}
-
-/// Builds a [`DebugSnapshot`] from the live `AppState`.
-///
-/// Shared by the `get_debug_snapshot` command and the bug reporter so the two
-/// can never capture divergent views of the session.
-pub(crate) async fn build_app_debug_snapshot(state: &Arc<AppState>) -> DebugSnapshot {
-    let world = state.world.lock().await;
-    let npc_manager = state.npc_manager.lock().await;
-    let events = state.debug_events.lock().await;
-    let game_events = state.game_events.lock().await;
-    let config = state.config.lock().await;
-
-    let call_log: Vec<parish_core::debug_snapshot::InferenceLogEntry> =
-        state.inference_log.lock().await.iter().cloned().collect();
-
-    let inference = InferenceDebug {
-        provider_name: config.provider_name.clone(),
-        model_name: config.model_name.clone(),
-        base_url: config.base_url.clone(),
-        cloud_provider: config.cloud_provider_name.clone(),
-        cloud_model: config.cloud_model_name.clone(),
-        has_queue: state.inference_queue.lock().await.is_some(),
-        reaction_req_id: parish_core::game_session::reaction_req_id_peek(),
-        improv_enabled: config.improv_enabled,
-        call_log,
-        categories: parish_core::debug_snapshot::build_inference_categories(&config),
-        configured_providers: parish_core::debug_snapshot::build_configured_providers(),
-        tier2_parse_failures_total: parish_core::npc::ticks::tier2_parse_failures_total(),
-    };
-
-    debug_snapshot::build_debug_snapshot(
-        &world,
-        &npc_manager,
-        &events,
-        &game_events,
-        &inference,
-        &AuthDebug::disabled(),
-    )
-}
-
-/// Returns the UI configuration from the loaded game mod.
-///
-/// The frontend uses this to set sidebar labels, accent colours, etc.
-#[tauri::command]
-pub async fn get_ui_config(
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<crate::UiConfigSnapshot, String> {
-    Ok(state.ui_config.clone())
-}
-
-/// Returns the latest provider-bootstrap status for the startup overlay.
-#[tauri::command]
-pub async fn get_setup_snapshot(
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<crate::SetupStatusSnapshot, String> {
-    Ok(state
-        .setup_status
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner())
-        .clone())
-}
-
-/// Toggles the desktop window between fullscreen and windowed mode.
-///
-/// Bound to F11 in the frontend (desktop only — the web build lets the
-/// browser handle native F11). Returns the resulting fullscreen state.
-#[tauri::command]
-pub async fn toggle_fullscreen(window: tauri::WebviewWindow) -> Result<bool, String> {
-    let target = !window.is_fullscreen().map_err(|e| e.to_string())?;
-    window.set_fullscreen(target).map_err(|e| e.to_string())?;
-    Ok(target)
-}
-
-// ── BYOK onboarding commands ─────────────────────────────────────────────────
-
-fn byok_ctx<'a>(state: &'a Arc<AppState>) -> parish_core::ipc::byok::ByokContext<'a> {
-    parish_core::ipc::byok::ByokContext {
-        config: &state.config,
-        inference_config: &state.inference_config,
-        inference_log: state.inference_log.clone(),
-        inference_file_log: state.inference_file_log.clone(),
-        slots: parish_core::game_loop::inference::InferenceSlots {
-            client: &state.client,
-            worker_handle: &state.worker_handle,
-            inference_queue: &state.inference_queue,
-        },
-        secrets: std::sync::Arc::clone(&state.secret_store),
-        user_config_dir: state.user_config_dir.as_path(),
-    }
-}
-
-/// Validates an unsaved provider/key combination by issuing a tiny live
-/// request. Used by the BYOK wizard before saving.
-#[tauri::command]
-pub async fn validate_provider_config(
-    args: parish_core::ipc::byok::ValidateProviderConfigArgs,
-) -> Result<parish_core::inference::validate::ValidationOutcome, String> {
-    Ok(parish_core::ipc::byok::handle_validate_provider_config(args).await)
-}
-
-/// Persists a BYOK config (keychain + parish.toml), updates GameConfig, and
-/// rebuilds the inference worker. Emits a fresh `setup-done` so the overlay
-/// dismisses.
-#[tauri::command]
-pub async fn set_provider_config(
-    args: parish_core::ipc::byok::SetProviderConfigArgs,
-    state: tauri::State<'_, Arc<AppState>>,
-    app: tauri::AppHandle,
-) -> Result<(), String> {
-    let state_arc = state.inner().clone();
-    parish_core::ipc::byok::handle_set_provider_config(args, byok_ctx(&state_arc))
-        .await
-        .map_err(|e| e.to_string())?;
-
-    {
-        let mut s = state_arc
-            .setup_status
-            .lock()
-            .unwrap_or_else(|p| p.into_inner());
-        s.clear_needs_onboarding();
-    }
-    crate::record_setup_done(&state_arc, true, String::new());
-    let _ = app.emit(
-        crate::events::EVENT_SETUP_DONE,
-        crate::events::SetupDonePayload {
-            success: true,
-            error: String::new(),
-        },
-    );
-    Ok(())
-}
-
-/// Returns the current effective provider config for the settings panel.
-/// Never returns the API key itself — only `has_api_key`/`has_env_key` flags.
-#[tauri::command]
-pub async fn get_provider_config(
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<parish_core::ipc::byok::GetProviderConfigResult, String> {
-    Ok(parish_core::ipc::byok::handle_get_provider_config(&state.config).await)
-}
-
-/// Wipes the keychain entry for the active provider and clears parish.toml.
-#[tauri::command]
-pub async fn clear_provider_config(state: tauri::State<'_, Arc<AppState>>) -> Result<(), String> {
-    let state_arc = state.inner().clone();
-    parish_core::ipc::byok::handle_clear_provider_config(byok_ctx(&state_arc))
-        .await
-        .map_err(|e| e.to_string())
-}
-
-/// Returns `{provider_id: has_env_key}` for every known provider so the BYOK
-/// wizard can show the env-detected hint on the picked provider, not just the
-/// current one.
-#[tauri::command]
-pub async fn list_byok_env_keys() -> Result<std::collections::BTreeMap<String, bool>, String> {
-    Ok(parish_core::ipc::byok::handle_list_env_keys())
-}
-
-/// Returns `{provider_id: {dialogue, simulation, intent, reaction}}` — the
-/// wizard uses this so its prefill matches what fill_missing_models_from_presets
-/// will actually install for the other tiers.
-#[tauri::command]
-pub async fn list_preset_models() -> Result<
-    std::collections::BTreeMap<String, Vec<parish_core::ipc::byok::ProviderPresetOption>>,
-    String,
-> {
-    Ok(parish_core::ipc::byok::handle_list_preset_models())
-}
-
-/// Returns the registry split into `featured` + `other` lists. The BYOK
-/// wizard renders directly from this — the source of truth is now the
-/// provider registry (builtins + mod-loaded), not a hand-curated TS
-/// constant.
-#[tauri::command]
-pub async fn list_available_providers() -> Result<
-    std::collections::HashMap<&'static str, Vec<parish_core::ipc::byok::ProviderInfo>>,
-    String,
-> {
-    Ok(parish_core::ipc::byok::handle_list_available_providers())
-}
-
-// ── #?: local-inference onboarding commands ────────────────────────────────
-
-/// Onboarding-choice payload returned to the SetupOverlay so it can render
-/// the right fork (local-recommended / local-low-mem / local-unavailable /
-/// configured).
-#[derive(serde::Serialize, serde::Deserialize)]
-pub struct OnboardingOptions {
-    /// The variant computed from platform + RAM + provider env.
-    pub choice: crate::setup::OnboardingChoice,
-    /// Unified memory in GB (macOS) or 0 elsewhere — feeds the "your Mac
-    /// has Xgb" copy on the local-recommended card.
-    pub ram_gb: u64,
-}
-
-/// Computes the onboarding fork to show the user on first run.
-///
-/// Pure read: no side effects on `AppState`. The SetupOverlay calls this
-/// once on mount to populate the fork UI; the same value is also persisted
-/// on `SetupStatusSnapshot.onboarding_choice` so reconnects after the
-/// EVENT_SETUP_NEEDS_ONBOARDING event still resolve.
-#[tauri::command]
-pub async fn get_onboarding_options(
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<OnboardingOptions, String> {
-    Ok(do_get_onboarding_options(state.inner()).await)
-}
-
-/// Internal helper shared with `mcp_bridge::onboarding_options`. Takes
-/// `&Arc<AppState>` so the MCP route can call it without the Tauri
-/// `State<'_, _>` wrapper.
-pub(crate) async fn do_get_onboarding_options(state: &Arc<AppState>) -> OnboardingOptions {
-    use parish_core::config::Provider;
-
-    let cfg = state.config.lock().await;
-    let provider_config = parish_core::config::ProviderConfig {
-        provider: Provider::from_str_loose(&cfg.provider_name).unwrap_or_default(),
-        api_key: cfg.api_key.clone(),
-        base_url: cfg.base_url.clone(),
-        model: Some(cfg.model_name.clone()),
-    };
-    drop(cfg);
-    let choice = crate::setup::onboarding_choice_for_platform(state, &provider_config);
-    let ram_gb = parish_core::config::unified_memory_bytes()
-        .map(|b| b / (1024 * 1024 * 1024))
-        .unwrap_or(0);
-    OnboardingOptions { choice, ram_gb }
-}
-
-/// Selection submitted from `LocalInferenceFork`. `two-slot` runs the
-/// 14B Dialogue + 1.5B small-slot loadout (recommended ≥16 GB);
-/// `small-only` runs Qwen2.5-1.5B-Instruct-4bit for every category on
-/// the same port (acceptable on <16 GB Macs at degraded quality).
-#[derive(serde::Deserialize)]
-pub struct LocalSetupArgs {
-    pub variant: String,
-}
-
-/// Persists local-inference config (provider=vllm-mlx + per-category
-/// overrides for the two-slot loadout), pre-downloads the required
-/// HuggingFace repos with progress reporting through the existing
-/// SetupOverlay surface, and rebuilds the inference worker.
-///
-/// On success, emits `setup-done` and clears the onboarding gate exactly
-/// like `set_provider_config` does for BYOK.
-#[tauri::command]
-pub async fn start_local_inference_setup(
-    args: LocalSetupArgs,
-    state: tauri::State<'_, Arc<AppState>>,
-    app: tauri::AppHandle,
-) -> Result<(), String> {
-    do_start_local_inference_setup(state.inner(), &app, args).await
-}
-
-/// Internal worker for the local-inference setup flow, shared with the MCP
-/// bridge route so an MCP client can drive the same wizard the desktop UI
-/// uses. Takes plain `&Arc<AppState>` + `&AppHandle` to sidestep the Tauri
-/// `State` wrapper that the `#[tauri::command]` shim provides.
-pub(crate) async fn do_start_local_inference_setup(
-    state: &Arc<AppState>,
-    app: &tauri::AppHandle,
-    args: LocalSetupArgs,
-) -> Result<(), String> {
-    use std::sync::atomic::Ordering;
-
-    // Idempotency guard — a second POST while the first is still running
-    // would race the bootstrap pipeline, double-spawn vllm-mlx serve, and
-    // produce undefined behaviour for the inference queue. Drop the
-    // duplicate with a busy error; the in-flight wizard keeps running.
-    // RAII guard restores the flag on every exit path.
-    struct WizardGuard<'a>(&'a std::sync::atomic::AtomicBool);
-    impl<'a> Drop for WizardGuard<'a> {
-        fn drop(&mut self) {
-            self.0.store(false, Ordering::Release);
-        }
-    }
-    if state
-        .wizard_in_flight
-        .compare_exchange(false, true, Ordering::AcqRel, Ordering::Acquire)
-        .is_err()
-    {
-        return Err("local-inference setup already in progress".to_string());
-    }
-    let _guard = WizardGuard(&state.wizard_in_flight);
-
-    // Error-path UX (#?) — every failing exit emits a `setup-done` with
-    // success=false + the error message, so the SetupOverlay drops out of
-    // the "Downloading…" spinner and the user sees what went wrong
-    // (network blip, disk full, vllm-mlx spawn failure). Without this the
-    // wizard hangs on the spinner forever and the user has to restart the
-    // app to see anything. The inner `result` is also returned to the
-    // caller so the MCP / Tauri shim can pass it back over the wire.
-    let inner = do_start_local_inference_setup_inner(state, app, args).await;
-    if let Err(ref e) = inner {
-        crate::record_setup_done(state, false, e.clone());
-        let _ = app.emit(
-            crate::events::EVENT_SETUP_DONE,
-            crate::events::SetupDonePayload {
-                success: false,
-                error: e.clone(),
-            },
-        );
-    }
-    inner
-}
-
-async fn do_start_local_inference_setup_inner(
-    state: &Arc<AppState>,
-    app: &tauri::AppHandle,
-    args: LocalSetupArgs,
-) -> Result<(), String> {
-    use parish_core::inference::hf_downloader::HfModelDownloader;
-    use parish_core::inference::setup::SetupProgress;
-    use std::sync::Arc as StdArc;
-
-    let state_arc = state.clone();
-
-    let two_slot = match args.variant.as_str() {
-        "two-slot" => true,
-        "small-only" => false,
-        other => return Err(format!("unknown variant: {other}")),
-    };
-
-    // Repos to fetch. The small slot is always present; the big slot only
-    // when the user picked two-slot (≥16 GB hosts).
-    let mut repos: Vec<&str> = vec!["mlx-community/Qwen2.5-1.5B-Instruct-4bit"];
-    if two_slot {
-        repos.insert(0, "mlx-community/Qwen2.5-14B-Instruct-4bit");
-    }
-
-    // Co-locate the HF cache under the user's app-config dir so a
-    // re-extracted bundle finds the existing models. `HF_HOME` env var
-    // gets set to this same path when we spawn vllm-mlx serve.
-    let hf_home = state_arc.user_config_dir.join("models");
-    std::fs::create_dir_all(&hf_home).map_err(|e| format!("mkdir {hf_home:?}: {e}"))?;
-
-    // `hf-hub` follows the Python convention: cache lives at
-    // `$HF_HOME/hub/...`. We pass the `hub` subdir directly to
-    // `with_cache_dir` (hf-hub treats it as the root for `models--…`).
-    let hf_hub = hf_home.join("hub");
-
-    let progress: StdArc<dyn SetupProgress> = StdArc::new(crate::TauriProgress::new(
-        app.clone(),
-        StdArc::clone(&state_arc),
-    ));
-
-    let downloader = HfModelDownloader::new(progress).with_cache_dir(hf_hub);
-    downloader
-        .download_models(&repos)
-        .await
-        .map_err(|e| format!("HF download failed: {e}"))?;
-
-    // Hand HF cache root + offline marker to the spawned vllm-mlx serve so
-    // it never re-checks the hub.
-    //
-    // SAFETY: set_var is unsafe on POSIX in multi-threaded contexts.
-    // The Tauri app is already running multi-threaded at this point, but
-    // we still need the env to be picked up by `VllmMlxProcess::ensure_running`
-    // when bootstrap runs. The variables are set before any worker spawn
-    // and the threads that read them do so within Command::env-style
-    // snapshots; this is the same approach used by parish_tauri::run() for
-    // VLLM_MLX_BIN.
-    unsafe {
-        std::env::set_var("PARISH_HF_HOME", &hf_home);
-    }
-
-    // Write provider config: two-slot puts Sim/Reaction/Intent on the
-    // small-slot port, Dialogue on the big-slot port. Small-only puts
-    // everything on the small-slot port.
-    let (provider_name, model_name, base_url) = if two_slot {
-        (
-            "vllm-mlx".to_string(),
-            "mlx-community/Qwen2.5-14B-Instruct-4bit".to_string(),
-            "http://localhost:8000".to_string(),
-        )
-    } else {
-        (
-            "vllm-mlx".to_string(),
-            "mlx-community/Qwen2.5-1.5B-Instruct-4bit".to_string(),
-            "http://localhost:8001".to_string(),
-        )
-    };
-
-    let mut category_overrides: std::collections::BTreeMap<
-        String,
-        parish_core::config::user_config::CategoryOverride,
-    > = std::collections::BTreeMap::new();
-    if two_slot {
-        // Intent stays on the small slot (1.5B) — fast classification
-        // doesn't need 14B and parse_intent's `Unknown` fallback handles
-        // any JSON parse failures (falls through to handle_npc_conversation
-        // as dialogue).
-        category_overrides.insert(
-            "intent".to_string(),
-            parish_core::config::user_config::CategoryOverride {
-                provider: Some("vllm-mlx".to_string()),
-                base_url: Some("http://localhost:8001".to_string()),
-                model: Some("mlx-community/Qwen2.5-1.5B-Instruct-4bit".to_string()),
-            },
-        );
-        // Sim + Reaction route to the simulator: the 1.5B can't reliably
-        // hold strict JSON for Tier 2 / Tier 3, and the resulting parse
-        // failures flooded the log every 5 game-seconds. The simulator
-        // returns valid `Tier2Response` / `Tier3Update` shapes (all
-        // `#[serde(default)]`), so ticks succeed as "uneventful" and the
-        // 1.5B is reserved for intent + the deterministic schedule logic
-        // handles NPC motion. The 14B big slot stays free for dialogue.
-        for cat_name in ["simulation", "reaction"] {
-            category_overrides.insert(
-                cat_name.to_string(),
-                parish_core::config::user_config::CategoryOverride {
-                    provider: Some("simulator".to_string()),
-                    base_url: None,
-                    model: None,
-                },
-            );
-        }
-    } else {
-        // small-only variant: 1.5B can't reliably hold the strict JSON
-        // schema Tier 2 (Simulation) and Tier 3 (Reaction) expect, and
-        // the resulting parse-failure storm both floods logs and starves
-        // the model slot Tier 1 needs for the dialogue stream. Route
-        // those categories to the in-process simulator so the
-        // living-world ticks stay quiet and every cycle of the 1.5B is
-        // spent on player-facing dialogue.
-        //
-        // Intent stays on vllm-mlx: parse_intent's failure path returns
-        // `Unknown` which falls through to `handle_npc_conversation`,
-        // i.e. a bad JSON parse still ends up routing player input as
-        // dialogue. The simulator's intent_json_for matches verb
-        // prefixes by `starts_with(mw)` without a word boundary, so
-        // "Good morning" gets classified as `Move`-to-"od morning" and
-        // the actual conversation never fires.
-        for cat_name in ["simulation", "reaction"] {
-            category_overrides.insert(
-                cat_name.to_string(),
-                parish_core::config::user_config::CategoryOverride {
-                    provider: Some("simulator".to_string()),
-                    base_url: None,
-                    model: None,
-                },
-            );
-        }
-    }
-
-    let set_args = parish_core::ipc::byok::SetProviderConfigArgs {
-        provider: provider_name,
-        base_url: Some(base_url),
-        model: Some(model_name),
-        api_key: None,
-        category_overrides,
-    };
-    parish_core::ipc::byok::handle_set_provider_config(set_args, byok_ctx(&state_arc))
-        .await
-        .map_err(|e| e.to_string())?;
-
-    // Clear the onboarding gate so the bootstrap path below doesn't
-    // bail out on the `onboarding_choice_for_platform` check.
-    {
-        let mut s = state_arc
-            .setup_status
-            .lock()
-            .unwrap_or_else(|p| p.into_inner());
-        s.clear_needs_onboarding();
-    }
-
-    // Run the same post-gate startup pipeline run() executes when the
-    // bootstrap completes cleanly on a returning user. Without this,
-    // the wizard would write the config + emit setup-done but the
-    // vllm-mlx serve process never spawns, no inference queue exists,
-    // and the world / autosave ticks never start — leaving the game
-    // visibly "ready" but functionally inert until a manual relaunch.
-    let (provider_config, _, _, _) = crate::provider_config_from_env(&state_arc.user_config_dir);
-    let inference_config_clone = state_arc.inference_config.clone();
-    let bootstrapped = crate::setup::bootstrap_inference_provider(
-        app,
-        &state_arc,
-        &provider_config,
-        &inference_config_clone,
-    )
-    .await;
-    if !bootstrapped {
-        return Err("inference bootstrap failed after wizard".to_string());
-    }
-    crate::setup::init_inference_queue(&state_arc).await;
-    crate::setup::init_persistence(app, &state_arc).await;
-    crate::setup::spawn_event_bus_fanin(&state_arc).await;
-    crate::setup::spawn_world_tick(app.clone(), Arc::clone(&state_arc));
-    crate::setup::spawn_inactivity_tick(app.clone(), Arc::clone(&state_arc));
-    crate::setup::spawn_debug_tick(app.clone(), Arc::clone(&state_arc));
-    crate::setup::spawn_autosave_tick(Arc::clone(&state_arc));
-    Ok(())
-}
-
-/// Processes player text input: classification → movement, look, or NPC conversation.
-///
-/// Movement and look results are resolved synchronously. NPC conversations
-/// submit an inference request and stream tokens back via `stream-token` events.
-#[tauri::command]
-pub async fn submit_input(
-    text: String,
-    addressed_to: Option<Vec<String>>,
-    state: tauri::State<'_, Arc<AppState>>,
-    app: tauri::AppHandle,
-) -> Result<(), String> {
-    do_submit_input(state.inner(), &app, text, addressed_to.unwrap_or_default()).await
-}
-
-/// Internal submit-input implementation shared with the MCP bridge.
-///
-/// Mirrors the Tauri command body but takes plain `&Arc<AppState>` /
-/// `&tauri::AppHandle` so the `mcp_bridge` Axum handler can drive the same
-/// dispatcher against the same live AppState the desktop window observes.
-pub(crate) async fn do_submit_input(
-    state: &Arc<AppState>,
-    app: &tauri::AppHandle,
-    text: String,
-    addressed_to: Vec<String>,
-) -> Result<(), String> {
-    let text = validate_input_text(&text)?;
-    if text.is_empty() {
-        return Ok(());
-    }
-    // #752 — cap addressed_to to prevent unbounded memory/allocation via the
-    // NPC-addressing chip list.  Max 10 entries; each name ≤ 100 chars.
-    validate_addressed_to(&addressed_to)?;
-
-    touch_player_activity(state).await;
-
-    // #9 — preempt any in-flight Tier 2 / Tier 3 sim call so the player's
-    // turn doesn't queue behind a 30 s constrained-decode batch. Cancel the
-    // current sim token (any sim call that snapshotted it will drop
-    // mid-stream and free the model slot) and install a fresh token so the
-    // next sim cycle has a live one to snapshot.
-    {
-        let mut sc = state.sim_cancel.lock().await;
-        sc.cancel();
-        *sc = tokio_util::sync::CancellationToken::new();
-    }
-
-    match classify_input(&text) {
-        InputResult::SystemCommand(cmd) => {
-            handle_system_command(cmd, state, app).await;
-        }
-        InputResult::GameInput(raw) => {
-            tracing::info!(input = %raw, "chat [player]");
-            // Emit the player's own text as a dialogue bubble only for actual dialogue
-            let player_msg = text_log("player", format!("> {}", raw));
-            let player_msg_id = player_msg.id.clone();
-            let _ = app.emit(EVENT_TEXT_LOG, player_msg);
-            let raw_for_reactions = raw.clone();
-            // Capture location before handle_game_input (which may move the player).
-            let reaction_location = state.world.lock().await.player_location;
-            handle_game_input(raw, addressed_to, state, app.clone()).await;
-            // Generate NPC reactions to the player's message in the background.
-            emit_npc_reactions(
-                &player_msg_id,
-                &raw_for_reactions,
-                reaction_location,
-                state,
-                app,
-            );
-        }
-    }
-
-    Ok(())
-}
-
-// ── #752 — addressed_to validation ───────────────────────────────────────────
-
-/// Validates and trims player free-text input for `submit_input`.
-///
-/// - Trims leading/trailing whitespace.
-/// - Returns `Ok(String)` (the trimmed text) for empty input — callers should
-///   short-circuit before calling this if they want to silently drop empties.
-/// - Returns `Err` when the trimmed length exceeds 2000 characters.
-pub fn validate_input_text(raw: &str) -> Result<String, String> {
-    let trimmed = raw.trim().to_string();
-    if trimmed.len() > 2000 {
-        return Err("Input too long (max 2000 characters).".to_string());
-    }
-    Ok(trimmed)
-}
-
-/// Maximum number of NPC chips a single `submit_input` may carry. Validated
-/// in [`validate_addressed_to`] and reused by `handle_game_input` to bound
-/// `Vec::with_capacity` calls (#933 — CodeQL `rust/uncontrolled-allocation-size`).
-pub(crate) const MAX_ADDRESSED_TO: usize = 10;
-
-/// Upper bound for the merged `addressed_to + mentions` target list. Sized
-/// generously above the realistic combined total — `addressed_to` is capped
-/// at [`MAX_ADDRESSED_TO`] and `mentions.names.len()` is bounded by NPCs in
-/// the world — so the allocation is guaranteed-small regardless of input.
-pub(crate) const MAX_TARGETS: usize = 64;
-
-/// Validates the `addressed_to` list from the `submit_input` command.
-///
-/// Rules (mode-parity with the server path in `parish-server`):
-/// - At most [`MAX_ADDRESSED_TO`] entries (prevents unbounded NPC-chip spam).
-/// - Each name is at most **100** characters.
-///
-/// Returns `Err(String)` with a user-visible message on any violation.
-pub fn validate_addressed_to(addressed_to: &[String]) -> Result<(), String> {
-    if addressed_to.len() > MAX_ADDRESSED_TO {
-        return Err(format!("Too many addressees (max {MAX_ADDRESSED_TO})."));
-    }
-    if addressed_to.iter().any(|name| name.len() > 100) {
-        return Err("Addressee name too long (max 100 characters).".to_string());
-    }
-    Ok(())
-}
-
-// ── Internal helpers ─────────────────────────────────────────────────────────
-
-/// Rebuilds the inference pipeline after a provider/key/client change.
-///
-/// Replaces the client and respawns the inference worker so subsequent
-/// NPC conversations use the new configuration.
-pub async fn rebuild_inference_inner(state: &Arc<AppState>, app: &tauri::AppHandle) {
-    let (provider_name, base_url, api_key) = {
-        let config = state.config.lock().await;
-        (
-            config.provider_name.clone(),
-            config.base_url.clone(),
-            config.api_key.clone(),
-        )
-    };
-
-    // Delegate to shared worker-lifecycle helper (#696).
-    let (_any_client, url_warning) = parish_core::game_loop::rebuild_inference_worker(
-        &provider_name,
-        &base_url,
-        api_key.as_deref(),
-        &state.inference_config,
-        state.inference_log.clone(),
-        state.inference_file_log.clone(),
-        parish_core::game_loop::inference::InferenceSlots {
-            client: &state.client,
-            worker_handle: &state.worker_handle,
-            inference_queue: &state.inference_queue,
-        },
-    )
-    .await;
-
-    // Surface URL warning via Tauri emit (Tauri-specific side effect).
-    if let Some(warn) = url_warning {
-        let _ = app.emit(
-            EVENT_TEXT_LOG,
-            TextLogPayload {
-                id: String::new(),
-                stream_turn_id: None,
-                source: "system".into(),
-                content: warn,
-                subtype: None,
-            },
-        );
-    }
-}
-
-async fn touch_player_activity(state: &Arc<AppState>) {
-    let mut conversation = state.conversation.lock().await;
-    let now = std::time::Instant::now();
-    conversation.last_player_activity = now;
-    conversation.last_spoken_at = now;
-}
-
-async fn emit_world_update(state: &Arc<AppState>, app: &tauri::AppHandle) {
-    let world = state.world.lock().await;
-    let npc_manager = state.npc_manager.lock().await;
-    let snapshot = get_world_snapshot_inner(&world, Some(&npc_manager), &state.pronunciations);
-    let _ = app.emit(EVENT_WORLD_UPDATE, snapshot);
-}
-
-/// Handles `/command` inputs.
-///
-/// Delegates to [`parish_core::game_loop::handle_system_command`] via the
-/// [`TauriCommandHost`] adapter (#696 slice 7).
-async fn handle_system_command(
-    cmd: parish_core::input::Command,
-    state: &Arc<AppState>,
-    app: &tauri::AppHandle,
-) {
-    use crate::command_host::TauriCommandHost;
-    use parish_core::game_loop::handle_system_command as shared_handle;
-
-    let host = TauriCommandHost::new(Arc::clone(state), app.clone());
-    shared_handle(&host, cmd).await;
-}
-
-/// Handles free-form game input: parses intent (with LLM fallback) then dispatches.
-///
-/// Takes plain `&Arc<AppState>` (not `tauri::State<...>`) so the body can be
-/// called from non-Tauri-extractor contexts — namely the `mcp_bridge` Axum
-/// handlers, which share the same live AppState as the desktop window. The
-/// Tauri callsite passes `state.inner()`.
-pub(crate) async fn handle_game_input(
-    raw: String,
-    addressed_to: Vec<String>,
-    state: &Arc<AppState>,
-    app: tauri::AppHandle,
-) {
-    // Resolve the intent client and model (Intent category override, or base).
-    let (client, model) = {
-        let config = state.config.lock().await;
-        let base_client = state.client.lock().await;
-        config.resolve_category_client(InferenceCategory::Intent, base_client.as_ref())
-    };
-
-    // Parse intent: tries local keywords first, then LLM for ambiguous input.
-    let intent = if let Some(client) = &client {
-        // Capture generation before releasing the lock so we can detect TOCTOU
-        // races on re-acquire (issue #283).
-        let gen_before = {
-            let mut world = state.world.lock().await;
-            world.clock.inference_pause();
-            world.tick_generation
-        };
-        let result = parse_intent(client, &raw, &model).await;
-        {
-            let mut world = state.world.lock().await;
-            world.clock.inference_resume();
-            let gen_after = world.tick_generation;
-            if gen_after != gen_before {
-                tracing::warn!(
-                    gen_before,
-                    gen_after,
-                    "World advanced during intent parse (TOCTOU #283) — \
-                     {} tick(s) elapsed; proceeding with parsed intent",
-                    gen_after.wrapping_sub(gen_before),
-                );
-                let _ = app.emit(
-                    crate::events::EVENT_TEXT_LOG,
-                    text_log(
-                        "system",
-                        "The world shifted while your words were in the air.",
-                    ),
-                );
-            }
-        }
-        result.ok()
-    } else {
-        // No client configured — use local keyword parsing only.
-        parish_core::input::parse_intent_local(&raw)
-    };
-
-    let is_move = intent
-        .as_ref()
-        .map(|i| matches!(i.intent, parish_core::input::IntentKind::Move))
-        .unwrap_or(false);
-    let is_look = intent
-        .as_ref()
-        .map(|i| matches!(i.intent, parish_core::input::IntentKind::Look))
-        .unwrap_or(false);
-    let is_talk = intent
-        .as_ref()
-        .map(|i| matches!(i.intent, parish_core::input::IntentKind::Talk))
-        .unwrap_or(false);
-    let move_target = intent
-        .as_ref()
-        .filter(|_i| is_move)
-        .and_then(|i| i.target.clone());
-    let talk_target = intent
-        .as_ref()
-        .filter(|_i| is_talk)
-        .and_then(|i| i.target.clone());
-
-    if is_move {
-        if let Some(target) = move_target {
-            handle_movement(&target, state, &app).await;
-        } else {
-            let _ = app.emit(
-                EVENT_TEXT_LOG,
-                TextLogPayload {
-                    id: String::new(),
-                    stream_turn_id: None,
-                    source: "system".into(),
-                    content: "And where would ye be off to?".to_string(),
-                    subtype: None,
-                },
-            );
-        }
-        return;
-    }
-
-    if is_look {
-        handle_look(state, &app).await;
-        return;
-    }
-
-    let mentions = {
-        let world = state.world.lock().await;
-        let npc_manager = state.npc_manager.lock().await;
-        parish_core::ipc::extract_npc_mentions(&raw, &world, &npc_manager)
-    };
-
-    // Chip selections (real names from the frontend) come first, then names
-    // detected in the player's text, then the LLM's single talk target when it
-    // supplied one. Deduping happens in `resolve_npc_targets` via
-    // `find_by_name`, which matches both real and display names.
-    // Pre-allocate at the fixed upper bound so the allocation argument is a
-    // constant — independent of any user-controlled input — and CodeQL's
-    // data-flow analyzer can see that.
-    let mut targets: Vec<String> = Vec::with_capacity(MAX_TARGETS);
-    for name in addressed_to {
-        if !targets.iter().any(|t| t == &name) {
-            targets.push(name);
-        }
-    }
-    for name in mentions.names {
-        if !targets.iter().any(|t| t == &name) {
-            targets.push(name);
-        }
-    }
-    if is_talk
-        && let Some(target) = talk_target
-        && !targets.iter().any(|t| t == &target)
-    {
-        targets.push(target);
-    }
-
-    handle_npc_conversation(mentions.remaining, targets, state, app).await;
-}
-
-/// Resolves movement to a named location using the shared movement pipeline.
-///
-/// Delegates all state mutation and message generation to
-/// [`parish_core::game_session::apply_movement`], then emits the returned
-/// effects to the frontend.
-async fn handle_movement(target: &str, state: &Arc<AppState>, app: &tauri::AppHandle) {
-    use parish_core::game_session::{
-        apply_movement, enrich_travel_encounter, roll_travel_encounter,
-    };
-
-    let transport = state.transport.default_mode().clone();
-
-    // Apply all movement state changes within a single lock scope to prevent
-    // TOCTOU races.
-    let (effects, rolled_encounter) = {
-        let mut world = state.world.lock().await;
-        let mut npc_manager = state.npc_manager.lock().await;
-        let effects = apply_movement(
-            &mut world,
-            &mut npc_manager,
-            &state.reaction_templates,
-            target,
-            &transport,
-        );
-        let rolled = if effects.world_changed {
-            let config = state.config.lock().await;
-            if !config.flags.is_disabled("travel-encounters") {
-                roll_travel_encounter(&world, &effects)
-            } else {
-                None
-            }
-        } else {
-            None
-        };
-        (effects, rolled)
-    };
-
-    // Resolve encounter text — LLM-enriched when a reaction client exists
-    // and the `travel-encounters-llm` flag is not explicitly disabled.
-    let encounter_line: Option<String> = if let Some(rolled) = rolled_encounter.as_ref() {
-        let llm_enabled = {
-            let cfg = state.config.lock().await;
-            !cfg.flags.is_disabled("travel-encounters-llm")
-        };
-        let (reaction_client, reaction_model) = if llm_enabled {
-            let config = state.config.lock().await;
-            let base_client = state.client.lock().await;
-            config.resolve_category_client(InferenceCategory::Reaction, base_client.as_ref())
-        } else {
-            (None, String::new())
-        };
-        let text = if let Some(client) = reaction_client.as_ref() {
-            enrich_travel_encounter(rolled, client, &reaction_model, 15).await
-        } else {
-            rolled.canned.text.clone()
-        };
-        let formatted = format!("  · {text}");
-        {
-            let mut world = state.world.lock().await;
-            world.log(formatted.clone());
-        }
-        Some(formatted)
-    } else {
-        None
-    };
-
-    // Emit travel-start animation payload first
-    if let Some(travel_payload) = &effects.travel_start {
-        let _ = app.emit(EVENT_TRAVEL_START, travel_payload);
-    }
-
-    // Emit all player-visible messages in order
-    for msg in &effects.messages {
-        tracing::info!(source = %msg.source, text = %msg.text.trim(), "chat");
-        let payload = match msg.subtype {
-            Some(st) => text_log_typed(msg.source, &msg.text, st),
-            None => text_log(msg.source, &msg.text),
-        };
-        let _ = app.emit(EVENT_TEXT_LOG, payload);
-    }
-
-    // Emit travel encounter line if one fired
-    if let Some(line) = encounter_line {
-        let _ = app.emit(EVENT_TEXT_LOG, text_log("system", &line));
-    }
-
-    // Emit NPC arrival reactions — stream gradually like normal NPC dialogue
-    if !effects.arrival_reactions.is_empty() {
-        use parish_core::game_session::stream_reaction_texts;
-
-        let (
-            all_npcs,
-            current_location_id,
-            loc_name,
-            tod,
-            weather,
-            introduced,
-            reaction_client,
-            reaction_model,
-        ) = {
-            let world = state.world.lock().await;
-            let npc_manager = state.npc_manager.lock().await;
-            let config = state.config.lock().await;
-            let base_client = state.client.lock().await;
-            let (rc, rm) =
-                config.resolve_category_client(InferenceCategory::Reaction, base_client.as_ref());
-            (
-                npc_manager.all_npcs().cloned().collect::<Vec<_>>(),
-                world.player_location,
-                world
-                    .current_location_data()
-                    .map(|d| d.name.clone())
-                    .unwrap_or_default(),
-                world.clock.time_of_day(),
-                world.weather.to_string(),
-                npc_manager.introduced_set(),
-                rc,
-                rm,
-            )
-        };
-
-        stream_reaction_texts(
-            &effects.arrival_reactions,
-            &all_npcs,
-            current_location_id,
-            &loc_name,
-            tod,
-            &weather,
-            &introduced,
-            reaction_client.as_ref(),
-            &reaction_model,
-            Some(&state.inference_log),
-            &state.language_settings,
-            |turn_id, npc_name| {
-                // Use `text_log_for_stream_turn` so the UI's streaming-
-                // placeholder guard recognises this entry and can finalise
-                // (remove) it when the per-turn `stream-turn-end` fires with
-                // no tokens — otherwise an empty bubble lingers in the chat
-                // (#984 follow-up: "blank NPC reply" reported on the
-                // `just demo 2 10` run).
-                let _ = app.emit(
-                    EVENT_TEXT_LOG,
-                    text_log_for_stream_turn(npc_name.to_string(), String::new(), turn_id),
-                );
-            },
-            |turn_id, source, batch| {
-                let _ = app.emit(
-                    EVENT_STREAM_TOKEN,
-                    StreamTokenPayload {
-                        token: batch.to_string(),
-                        turn_id,
-                        source: source.to_string(),
-                        // Arrival reactions have no reconnect-resume contract
-                        // (desktop transport never disconnects); `None` keeps
-                        // the wire shape unchanged (#1164).
-                        message_id: None,
-                    },
-                );
-            },
-            |turn_id| {
-                let _ = app.emit(EVENT_STREAM_TURN_END, StreamTurnEndPayload { turn_id });
-            },
-        )
-        .await;
-
-        // Finalise the streaming state so the frontend marks the last entry done.
-        let _ = app.emit(EVENT_STREAM_END, StreamEndPayload { hints: vec![] });
-    }
-
-    // Record tier transitions in the debug event log
-    if !effects.tier_transitions.is_empty() {
-        let ts = debug_event_timestamp(state).await;
-        let mut debug_events = state.debug_events.lock().await;
-        for tt in &effects.tier_transitions {
-            if debug_events.len() >= crate::DEBUG_EVENT_CAPACITY {
-                debug_events.pop_front();
-            }
-            let direction = if tt.promoted { "promoted" } else { "demoted" };
-            debug_events.push_back(DebugEvent {
-                timestamp: ts.clone(),
-                category: "tier".to_string(),
-                message: format!(
-                    "{} {} {:?} → {:?}",
-                    tt.npc_name, direction, tt.old_tier, tt.new_tier,
-                ),
-            });
-        }
-    }
-
-    // Emit updated world snapshot after a successful move
-    if effects.world_changed {
-        let current_location = {
-            let world = state.world.lock().await;
-            world.player_location
-        };
-        let mut conversation = state.conversation.lock().await;
-        conversation.sync_location(current_location);
-        conversation.last_spoken_at = std::time::Instant::now();
-        drop(conversation);
-
-        let world = state.world.lock().await;
-        let npc_manager = state.npc_manager.lock().await;
-        let snapshot = get_world_snapshot_inner(&world, Some(&npc_manager), &state.pronunciations);
-        let _ = app.emit(EVENT_WORLD_UPDATE, snapshot);
-    }
-}
-
-/// Renders the current location description and exits.
-async fn handle_look(state: &Arc<AppState>, app: &tauri::AppHandle) {
-    let world = state.world.lock().await;
-    let npc_manager = state.npc_manager.lock().await;
-    let transport = state.transport.default_mode();
-    let text = parish_core::ipc::render_look_text(
-        &world,
-        &npc_manager,
-        transport.speed_m_per_s,
-        &transport.label,
-        false,
-    );
-    let _ = app.emit(
-        EVENT_TEXT_LOG,
-        TextLogPayload {
-            id: String::new(),
-            stream_turn_id: None,
-            source: "system".into(),
-            content: text,
-            subtype: None,
-        },
-    );
-}
-
-/// Helper: sets `conversation_in_progress` on the conversation mutex.
-///
-/// Only used in unit tests; production code uses the `GameLoopContext`-based
-/// shared orchestration which manages this flag internally.
-#[cfg(test)]
-async fn set_conversation_running(state: &Arc<AppState>, running: bool) {
-    let mut conversation = state.conversation.lock().await;
-    conversation.conversation_in_progress = running;
-}
-
-/// Routes input to one or more NPCs at the player's location, or shows an idle message.
-///
-/// Delegates to [`parish_core::game_loop::handle_npc_conversation`] for all
-/// shared logic (#696), then emits a world-update snapshot when inference
-/// finishes.
-async fn handle_npc_conversation(
-    raw: String,
-    target_names: Vec<String>,
-    state: &Arc<AppState>,
-    app: tauri::AppHandle,
-) {
-    let emitter: std::sync::Arc<dyn parish_core::ipc::EventEmitter> =
-        std::sync::Arc::new(crate::events::TauriEmitter::new(app.clone()));
-    let ctx = parish_core::game_loop::GameLoopContext {
-        world: &state.world,
-        npc_manager: &state.npc_manager,
-        config: &state.config,
-        conversation: &state.conversation,
-        inference_queue: &state.inference_queue,
-        emitter: std::sync::Arc::clone(&emitter),
-        inference_config: &state.inference_config,
-        pronunciations: &state.pronunciations,
-        client: &state.client,
-        cloud_client: &state.cloud_client,
-        language: state.language_settings.clone(),
-        inference_failure_messages: &state.inference_failure_messages,
-        idle_messages: &state.idle_messages,
-    };
-
-    let app_for_loading = app.clone();
-    let spawn_loading = move || {
-        let cancel = tokio_util::sync::CancellationToken::new();
-        crate::events::spawn_loading_animation(app_for_loading.clone(), cancel.clone());
-        Some(cancel)
-    };
-
-    emit_world_update(state, &app).await;
-    parish_core::game_loop::handle_npc_conversation(&ctx, raw, target_names, spawn_loading).await;
-    emit_world_update(state, &app).await;
-}
-
-/// Delegates to [`parish_core::game_loop::run_idle_banter`] for all shared
-/// logic (#696), then emits a world-update snapshot when the sequence ends.
-async fn run_idle_banter(state: &Arc<AppState>, app: &tauri::AppHandle) {
-    let emitter: std::sync::Arc<dyn parish_core::ipc::EventEmitter> =
-        std::sync::Arc::new(crate::events::TauriEmitter::new(app.clone()));
-    let ctx = parish_core::game_loop::GameLoopContext {
-        world: &state.world,
-        npc_manager: &state.npc_manager,
-        config: &state.config,
-        conversation: &state.conversation,
-        inference_queue: &state.inference_queue,
-        emitter: std::sync::Arc::clone(&emitter),
-        inference_config: &state.inference_config,
-        pronunciations: &state.pronunciations,
-        client: &state.client,
-        cloud_client: &state.cloud_client,
-        language: state.language_settings.clone(),
-        inference_failure_messages: &state.inference_failure_messages,
-        idle_messages: &state.idle_messages,
-    };
-
-    emit_world_update(state, app).await;
-    // Idle banter spawns no loading animation.
-    parish_core::game_loop::run_idle_banter(&ctx, || None).await;
-    emit_world_update(state, app).await;
-}
-
-pub(crate) async fn tick_inactivity(state: &Arc<AppState>, app: &tauri::AppHandle) {
-    let (last_player_activity, last_spoken_at, running, idle_after, auto_pause_after) = {
-        let conversation = state.conversation.lock().await;
-        let config = state.config.lock().await;
-        (
-            conversation.last_player_activity,
-            conversation.last_spoken_at,
-            conversation.conversation_in_progress,
-            config.idle_banter_after_secs,
-            config.auto_pause_after_secs,
-        )
-    };
-
-    if running {
-        return;
-    }
-
-    let world_state = {
-        let world = state.world.lock().await;
-        (
-            world.clock.is_paused(),
-            world.clock.is_inference_paused(),
-            world.player_location,
-        )
-    };
-
-    if world_state.0 || world_state.1 {
-        return;
-    }
-
-    {
-        let mut conversation = state.conversation.lock().await;
-        conversation.sync_location(world_state.2);
-    }
-
-    let now = std::time::Instant::now();
-    let player_idle = now.duration_since(last_player_activity).as_secs();
-    let speech_idle = now.duration_since(last_spoken_at).as_secs();
-
-    if player_idle >= auto_pause_after {
-        {
-            let mut world = state.world.lock().await;
-            if world.clock.is_paused() || world.clock.is_inference_paused() {
-                return;
-            }
-            world.clock.pause();
-        }
-        let _ = app.emit(
-            EVENT_TEXT_LOG,
-            TextLogPayload {
-                id: String::new(),
-                stream_turn_id: None,
-                source: "system".into(),
-                content:
-                    "The parish falls quiet after a full minute of silence. Time is now paused."
-                        .to_string(),
-                subtype: None,
-            },
-        );
-        emit_world_update(state, app).await;
-        let mut conversation = state.conversation.lock().await;
-        conversation.last_spoken_at = now;
-        return;
-    }
-
-    if player_idle >= idle_after && speech_idle >= idle_after {
-        run_idle_banter(state, app).await;
-    }
-}
-
-// ── Persistence commands ────────────────────────────────────────────────────
-
-use parish_core::persistence::Database;
-use parish_core::persistence::picker::{SaveFileInfo, discover_saves, new_save_path};
-use parish_core::persistence::snapshot::GameSnapshot;
-
-/// Returns the list of save files with branch metadata.
-#[tauri::command]
-pub async fn discover_save_files(
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<Vec<SaveFileInfo>, String> {
-    let world = state.world.lock().await;
-    let saves = discover_saves(&state.saves_dir, &world.graph);
-    for s in &saves {
-        tracing::info!(
-            "Save file: {} — {} branches: {:?}",
-            s.filename,
-            s.branches.len(),
-            s.branches.iter().map(|b| &b.name).collect::<Vec<_>>()
-        );
-    }
-    Ok(saves)
-}
-
-/// Saves the current game state to the active save file and branch.
-///
-/// If no save file is active, creates a new one.
-#[tauri::command]
-pub async fn save_game(state: tauri::State<'_, Arc<AppState>>) -> Result<String, String> {
-    do_save_game(&state).await
-}
-
-/// Internal save implementation — delegates to the shared canonical impl (#696).
-pub(crate) async fn do_save_game(state: &Arc<AppState>) -> Result<String, String> {
-    parish_core::game_loop::do_save_game(
-        &state.world,
-        &state.npc_manager,
-        &state.save_path,
-        &state.current_branch_id,
-        &state.current_branch_name,
-        &state.saves_dir,
-    )
-    .await
-}
-
-/// Loads a branch from a save file, restoring world and NPC state.
-#[tauri::command]
-pub async fn load_branch(
-    file_path: String,
-    branch_id: i64,
-    state: tauri::State<'_, Arc<AppState>>,
-    app: tauri::AppHandle,
-) -> Result<(), String> {
-    do_load_branch(state.inner(), &app, file_path, branch_id).await
-}
-
-/// Internal load-branch implementation shared with the MCP bridge.
-///
-/// Takes plain `&Arc<AppState>` / `&tauri::AppHandle` so it can be called
-/// from non-Tauri-extractor contexts (e.g. `mcp_bridge::load_branch_route`).
-pub async fn do_load_branch(
-    state: &Arc<AppState>,
-    app: &tauri::AppHandle,
-    file_path: String,
-    branch_id: i64,
-) -> Result<(), String> {
-    use parish_core::persistence::SaveFileLock;
-
-    let path = std::path::PathBuf::from(&file_path);
-
-    // If switching to a different save file, acquire a new lock first.
-    let current_path = state.save_path.lock().await.clone();
-    let switching_files = current_path.as_ref() != Some(&path);
-
-    if switching_files {
-        let lock = SaveFileLock::try_acquire(&path)
-            .ok_or_else(|| "This save file is in use by another instance.".to_string())?;
-        // Release old lock and store new one.
-        *state.save_lock.lock().await = Some(lock);
-    }
-
-    let path_clone = path.clone();
-    let (snapshot, branch_name) =
-        tokio::task::spawn_blocking(move || -> Result<(GameSnapshot, String), String> {
-            let db = Database::open(&path_clone).map_err(|e| e.to_string())?;
-            let (_, snapshot) = db
-                .load_latest_snapshot(branch_id)
-                .map_err(|e| e.to_string())?
-                .ok_or_else(|| "No snapshots found on this branch.".to_string())?;
-            let branches = db.list_branches().map_err(|e| e.to_string())?;
-            let branch_name = branches
-                .iter()
-                .find(|b| b.id == branch_id)
-                .map(|b| b.name.clone())
-                .unwrap_or_else(|| "unknown".to_string());
-            Ok((snapshot, branch_name))
-        })
-        .await
-        .map_err(|e| e.to_string())??;
-
-    // Restore state
-    let mut world = state.world.lock().await;
-    let mut npc_manager = state.npc_manager.lock().await;
-    snapshot.restore(&mut world, &mut npc_manager);
-    npc_manager.assign_tiers(&world, &[]);
-
-    // Update save tracking
-    let filename = path
-        .file_name()
-        .map(|n| n.to_string_lossy().to_string())
-        .unwrap_or_default();
-
-    // Emit updated state to frontend (compute name hints before dropping locks)
-    let ws = get_world_snapshot_inner(&world, Some(&npc_manager), &state.pronunciations);
-    drop(npc_manager);
-    let _ = app.emit(EVENT_WORLD_UPDATE, ws);
-    let _ = app.emit(
-        EVENT_TEXT_LOG,
-        TextLogPayload {
-            id: String::new(),
-            stream_turn_id: None,
-            source: "system".into(),
-            content: format!("Loaded {} (branch: {}).", filename, branch_name),
-            subtype: None,
-        },
-    );
-
-    drop(world);
-
-    // Update persistence tracking
-    *state.save_path.lock().await = Some(path);
-    *state.current_branch_id.lock().await = Some(branch_id);
-    *state.current_branch_name.lock().await = Some(branch_name);
-
-    Ok(())
-}
-
-/// Creates a new branch forked from a specified parent branch.
-#[tauri::command]
-pub async fn create_branch(
-    name: String,
-    parent_branch_id: i64,
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<String, String> {
-    do_create_branch(&state, &name, parent_branch_id).await
-}
-
-/// Internal fork implementation shared by the command and /fork handler.
-pub async fn do_create_branch(
-    state: &Arc<AppState>,
-    name: &str,
-    parent_branch_id: i64,
-) -> Result<String, String> {
-    // #1196 — validate before touching locks or the database.
-    parish_core::input::validate_branch_name(name)
-        .map_err(|e| format!("Invalid branch name: {e}"))?;
-
-    let db_path = {
-        let guard = state.save_path.lock().await;
-        guard
-            .as_ref()
-            .ok_or("No active save file. Use /save first.")?
-            .clone()
-    };
-
-    tracing::info!(
-        "Creating branch '{}' with parent {} in {:?}",
-        name,
-        parent_branch_id,
-        db_path
-    );
-
-    // Capture snapshot before spawn_blocking so the tokio locks are not held across it.
-    let world = state.world.lock().await;
-    let npc_manager = state.npc_manager.lock().await;
-    let snapshot = GameSnapshot::capture(&world, &npc_manager);
-    drop(npc_manager);
-    drop(world);
-
-    let name_owned = name.to_string();
-    let new_id = tokio::task::spawn_blocking(move || -> Result<i64, String> {
-        let db = Database::open(&db_path).map_err(|e| e.to_string())?;
-        let new_id = db
-            .create_branch(&name_owned, Some(parent_branch_id))
-            .map_err(|e| {
-                tracing::error!("create_branch failed: {}", e);
-                e.to_string()
-            })?;
-        tracing::info!("Branch '{}' created with id {}", name_owned, new_id);
-        db.save_snapshot(new_id, &snapshot)
-            .map_err(|e| e.to_string())?;
-        tracing::info!("Snapshot saved to branch '{}'", name_owned);
-        Ok(new_id)
-    })
-    .await
-    .map_err(|e| e.to_string())??;
-
-    // Switch to the new branch
-    *state.current_branch_id.lock().await = Some(new_id);
-    *state.current_branch_name.lock().await = Some(name.to_string());
-
-    Ok(format!("Created new branch '{}'.", name))
-}
-
-/// Creates a new save file and saves the current state.
-#[tauri::command]
-pub async fn new_save_file(state: tauri::State<'_, Arc<AppState>>) -> Result<(), String> {
-    use parish_core::persistence::SaveFileLock;
-
-    let path = new_save_path(&state.saves_dir);
-
-    // Acquire lock on the new save file, releasing any previous lock.
-    let lock = SaveFileLock::try_acquire(&path)
-        .ok_or_else(|| "Could not lock the new save file.".to_string())?;
-    *state.save_lock.lock().await = Some(lock);
-
-    let world = state.world.lock().await;
-    let npc_manager = state.npc_manager.lock().await;
-    let snapshot = GameSnapshot::capture(&world, &npc_manager);
-    drop(npc_manager);
-    drop(world);
-
-    let path_clone = path.clone();
-    let branch_id = tokio::task::spawn_blocking(move || -> Result<i64, String> {
-        let db = Database::open(&path_clone).map_err(|e| e.to_string())?;
-        let branch = db
-            .find_branch("main")
-            .map_err(|e| e.to_string())?
-            .ok_or("Failed to create main branch")?;
-        db.save_snapshot(branch.id, &snapshot)
-            .map_err(|e| e.to_string())?;
-        Ok(branch.id)
-    })
-    .await
-    .map_err(|e| e.to_string())??;
-
-    *state.save_path.lock().await = Some(path);
-    *state.current_branch_id.lock().await = Some(branch_id);
-    *state.current_branch_name.lock().await = Some("main".to_string());
-
-    Ok(())
-}
-
-/// Internal helper that reloads world/NPCs and creates a fresh save file.
-///
-/// Called both by the `new_game` Tauri command and the `CommandEffect::NewGame`
-/// handler.  Delegates to the shared `parish_core::game_loop::do_new_game` (#696).
-pub async fn do_new_game(state: &Arc<AppState>, app: &tauri::AppHandle) -> Result<(), String> {
-    use parish_core::game_loop::{NewGameParams, do_new_game as core_do_new_game};
-
-    // Rule 9 (#1197): use the mod resolved once at startup and stored on
-    // AppState, not a per-call cwd-walk via find_default_mod().
-    let emitter = crate::events::TauriEmitter::new(app.clone());
-    core_do_new_game(NewGameParams {
-        world: &state.world,
-        npc_manager: &state.npc_manager,
-        conversation: &state.conversation,
-        save_path: &state.save_path,
-        current_branch_id: &state.current_branch_id,
-        current_branch_name: &state.current_branch_name,
-        saves_dir: &state.saves_dir,
-        game_mod: state.game_mod.as_ref(),
-        data_dir: &state.data_dir,
-        pronunciations: &state.pronunciations,
-        default_transport: state.transport.default_mode(),
-        emitter: &emitter,
-    })
-    .await
-}
-
-/// Starts a brand new game: reloads world and NPCs from data files,
-/// creates a new save file, and saves the fresh initial state.
-#[tauri::command]
-pub async fn new_game(
-    state: tauri::State<'_, Arc<AppState>>,
-    app: tauri::AppHandle,
-) -> Result<(), String> {
-    do_new_game(&state, &app).await?;
-    let _ = app.emit(
-        EVENT_TEXT_LOG,
-        TextLogPayload {
-            id: String::new(),
-            stream_turn_id: None,
-            source: "system".into(),
-            content: "A new chapter begins in the parish...".to_string(),
-            subtype: None,
-        },
-    );
-    Ok(())
-}
-
-/// Returns the current save state for display in the StatusBar.
-#[tauri::command]
-pub async fn get_save_state(state: tauri::State<'_, Arc<AppState>>) -> Result<SaveState, String> {
-    let save_path = state.save_path.lock().await;
-    let branch_id = state.current_branch_id.lock().await;
-    let branch_name = state.current_branch_name.lock().await;
-
-    Ok(SaveState {
-        filename: save_path
-            .as_ref()
-            .and_then(|p| p.file_name())
-            .map(|n| n.to_string_lossy().to_string()),
-        branch_id: *branch_id,
-        branch_name: branch_name.clone(),
-    })
-}
-
-// ── Bug reporting ─────────────────────────────────────────────────────────────
-
-/// Files a bug report — bundles a screenshot, recent logs, and current game
-/// state into a GitHub issue (or, in dry-run / no-token mode, a bundle on
-/// disk). Shared with the MCP bridge's `/api/submit-bug-report` route.
-#[tauri::command]
-pub async fn submit_bug_report(
-    title: String,
-    description: Option<String>,
-    screenshot_data_url: Option<String>,
-    context: Option<parish_core::ipc::BugContext>,
-    state: tauri::State<'_, Arc<AppState>>,
-    app: tauri::AppHandle,
-) -> Result<parish_core::ipc::BugReportResult, String> {
-    let request = parish_core::ipc::BugReportRequest {
-        title,
-        description: description.unwrap_or_default(),
-        screenshot_data_url,
-        context,
-    };
-    do_submit_bug_report(&state, &app, request).await
-}
-
-/// Shared bug-report implementation (Tauri command + MCP bridge route).
-///
-/// Gathers a world + debug snapshot and a save summary from the live
-/// `AppState`, resolves the screenshot (decoding the frontend-supplied data
-/// URL, or triggering a live `request-screenshot` round-trip when absent),
-/// then delegates the GitHub work to `parish_core::ipc::bug_report`.
-pub(crate) async fn do_submit_bug_report(
-    state: &Arc<AppState>,
-    app: &tauri::AppHandle,
-    request: parish_core::ipc::BugReportRequest,
-) -> Result<parish_core::ipc::BugReportResult, String> {
-    use parish_core::ipc::bug_report;
-
-    // Feature flag (default-on kill switch, per AGENTS.md §6).
-    if state.config.lock().await.flags.is_disabled("bug-report") {
-        return Err(bug_report::BugReportError::Disabled.to_string());
-    }
-
-    let world_snapshot = {
-        let world = state.world.lock().await;
-        parish_core::ipc::snapshot_from_world(&world)
-    };
-    let debug = build_app_debug_snapshot(state).await;
-    let save_summary = {
-        let branch_id = *state.current_branch_id.lock().await;
-        let branch_name = state.current_branch_name.lock().await.clone();
-        match (branch_id, branch_name) {
-            (Some(id), Some(name)) => Some(format!("branch {id}: {name}")),
-            (Some(id), None) => Some(format!("branch {id}")),
-            (None, Some(name)) => Some(name),
-            (None, None) => None,
-        }
-    };
-    let report_state =
-        bug_report::BugReportState::from_snapshots(&world_snapshot, &debug, save_summary);
-
-    // Resolve screenshot bytes: prefer the frontend-supplied data URL, else
-    // trigger a live capture and read it back. A failed/timed-out capture is
-    // non-fatal — the report is still filed without an image.
-    let screenshot_png: Option<Vec<u8>> = match &request.screenshot_data_url {
-        Some(data_url) => Some(bug_report::decode_data_url(data_url).map_err(|e| e.to_string())?),
-        None => match do_take_screenshot(state, app).await {
-            Ok(info) => tokio::fs::read(&info.path).await.ok(),
-            Err(_) => None,
-        },
-    };
-
-    let cfg = bug_report::GitHubBugConfig::from_env_async().await;
-    let bundle_root = state.saves_dir.join("bug-reports");
-    let http = reqwest::Client::new();
-    bug_report::create_bug_report(
-        &http,
-        &cfg,
-        &request,
-        &report_state,
-        screenshot_png.as_deref(),
-        &bundle_root,
-    )
-    .await
-    .map_err(|e| e.to_string())
-}
-
-// ── Screenshot capture (player-triggered, MCP-readable) ──────────────────────
-
-/// Metadata describing the most-recently-saved screenshot.
-///
-/// The image is written to `<saves_dir>/screenshots/parish-<ISO-timestamp>.png`
-/// by the Svelte frontend (which calls `html-to-image` and posts the resulting
-/// `data:image/png;base64,...` URL to the `save_screenshot` command). Once
-/// written, the absolute path is cached on `AppState::latest_screenshot_path`
-/// so the `parish_latest_screenshot` MCP tool can report it without scanning
-/// the directory.
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct ScreenshotInfo {
-    /// Absolute filesystem path to the PNG.
-    pub path: String,
-    /// ISO-8601 UTC timestamp the file was written (`YYYY-MM-DDTHH:MM:SSZ`).
-    pub taken_at: String,
-    /// Size of the PNG payload in bytes.
-    pub size_bytes: u64,
-}
-
-/// Decodes a `data:image/png;base64,...` URL into the raw PNG byte payload.
-///
-/// Returns `Err` if the URL is malformed, has the wrong MIME type, or the
-/// base64 segment fails to decode.
-pub fn decode_data_url_png(data_url: &str) -> Result<Vec<u8>, String> {
-    use base64::Engine as _;
-    const PREFIX: &str = "data:image/png;base64,";
-    let b64 = data_url
-        .strip_prefix(PREFIX)
-        .ok_or_else(|| format!("expected data URL to start with `{PREFIX}`"))?;
-    base64::engine::general_purpose::STANDARD
-        .decode(b64.as_bytes())
-        .map_err(|e| format!("base64 decode failed: {e}"))
-}
-
-/// Writes the decoded PNG bytes under `<saves_dir>/screenshots/parish-<ISO>.png`
-/// and returns the [`ScreenshotInfo`] metadata for the newly created file.
-///
-/// Pure on `(saves_dir, png_bytes, now)` — no AppState, no Tauri handle — so
-/// it can be unit-tested in isolation. The `now` callback returns the UTC
-/// timestamp used both in the filename and the response (it is parameterised
-/// so tests can pin the value).
-pub fn write_screenshot_to_disk(
-    saves_dir: &std::path::Path,
-    png_bytes: &[u8],
-    now: chrono::DateTime<chrono::Utc>,
-) -> Result<ScreenshotInfo, String> {
-    let dir = saves_dir.join("screenshots");
-    std::fs::create_dir_all(&dir)
-        .map_err(|e| format!("failed to create {}: {e}", dir.display()))?;
-
-    // Filenames must be filesystem-safe on every platform we support, so use
-    // `-` instead of `:` in the time component. `format!("{:?}", ts)` would
-    // include sub-second precision plus the trailing `Z`; we keep the stem
-    // tidy by formatting with the second-precision strftime template.
-    let stem = now.format("parish-%Y-%m-%dT%H-%M-%SZ").to_string();
-    let path = dir.join(format!("{stem}.png"));
-
-    std::fs::write(&path, png_bytes)
-        .map_err(|e| format!("failed to write {}: {e}", path.display()))?;
-
-    let size_bytes = png_bytes.len() as u64;
-    let path_string = path.to_string_lossy().to_string();
-    let taken_at = now.format("%Y-%m-%dT%H:%M:%SZ").to_string();
-    Ok(ScreenshotInfo {
-        path: path_string,
-        taken_at,
-        size_bytes,
-    })
-}
-
-/// Internal save-screenshot implementation shared with the MCP bridge / web
-/// route. Decodes the `data_url`, writes the PNG, and updates
-/// [`AppState::latest_screenshot_path`].
-pub(crate) async fn do_save_screenshot(
-    state: &Arc<AppState>,
-    data_url: String,
-) -> Result<ScreenshotInfo, String> {
-    let bytes = decode_data_url_png(&data_url)?;
-    let info = write_screenshot_to_disk(&state.saves_dir, &bytes, chrono::Utc::now())?;
-    *state.latest_screenshot_path.lock().await = Some(std::path::PathBuf::from(&info.path));
-    Ok(info)
-}
-
-/// Internal latest-screenshot reader shared with the MCP bridge / web route.
-///
-/// Re-stat`s the file each call so a screenshot deleted out from under the
-/// session is reported as missing rather than reused indefinitely.
-pub(crate) async fn do_get_latest_screenshot(
-    state: &Arc<AppState>,
-) -> Result<Option<ScreenshotInfo>, String> {
-    let Some(path) = state.latest_screenshot_path.lock().await.clone() else {
-        return Ok(None);
-    };
-    // Use tokio::fs::metadata so the stat call doesn't block the async
-    // executor under load (this handler may be invoked from the MCP bridge
-    // while other Tokio tasks are waiting on the same worker thread).
-    let metadata = match tokio::fs::metadata(&path).await {
-        Ok(m) => m,
-        Err(_) => return Ok(None),
-    };
-    let modified = metadata
-        .modified()
-        .map_err(|e| format!("stat({}): {e}", path.display()))?;
-    let taken_at = chrono::DateTime::<chrono::Utc>::from(modified)
-        .format("%Y-%m-%dT%H:%M:%SZ")
-        .to_string();
-    Ok(Some(ScreenshotInfo {
-        path: path.to_string_lossy().to_string(),
-        taken_at,
-        size_bytes: metadata.len(),
-    }))
-}
-
-/// Persists a screenshot captured by the frontend.
-///
-/// `data_url` is a `data:image/png;base64,...` string produced by
-/// `html-to-image` in the Svelte UI. The PNG is decoded and written to
-/// `<saves_dir>/screenshots/parish-<ISO-timestamp>.png`; the resulting path
-/// is cached on [`AppState::latest_screenshot_path`] so the MCP
-/// `parish_latest_screenshot` tool can read it back without rescanning.
-#[tauri::command]
-pub async fn save_screenshot(
-    data_url: String,
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<ScreenshotInfo, String> {
-    do_save_screenshot(&state, data_url).await
-}
-
-/// Reads metadata for the most recently captured screenshot, if any.
-///
-/// Returns `Ok(None)` when no screenshot has been captured this session, or
-/// when the cached path no longer exists on disk.
-#[tauri::command]
-pub async fn get_latest_screenshot(
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<Option<ScreenshotInfo>, String> {
-    do_get_latest_screenshot(&state).await
-}
-
-/// Agent-triggered screenshot capture. Registered as a Tauri command for
-/// wiring-parity with the `/api/take-screenshot` bridge route.
-///
-/// In Tauri mode this is not invoked directly by the frontend — use the MCP
-/// tool `parish_take_screenshot` instead. The bridge handler also calls
-/// `do_take_screenshot` directly for a shared implementation.
-#[tauri::command]
-pub async fn take_screenshot(
-    state: tauri::State<'_, Arc<AppState>>,
-    app: tauri::AppHandle,
-) -> Result<ScreenshotInfo, String> {
-    do_take_screenshot(&state, &app).await
-}
-
-/// Capture deadline for the agent screenshot round-trip.
-///
-/// Reads `PARISH_SCREENSHOT_TIMEOUT_SECS` (whole seconds); defaults to 45 and
-/// ignores unparseable or zero values. The full `.app-shell` `html-to-image`
-/// pass can exceed the old hardcoded 15 s under live local-inference load
-/// (#1160), so the deadline is configurable and longer by default.
-fn screenshot_timeout() -> std::time::Duration {
-    let secs = std::env::var("PARISH_SCREENSHOT_TIMEOUT_SECS")
-        .ok()
-        .and_then(|v| v.trim().parse::<u64>().ok())
-        .filter(|&n| n > 0)
-        .unwrap_or(45);
-    std::time::Duration::from_secs(secs)
-}
-
-/// Decides whether a late-completing capture can still be returned after the
-/// oneshot deadline expired.
-///
-/// The frontend may finish `html-to-image` and write the PNG just after the
-/// backend gave up waiting; without this, that image is orphaned and the
-/// caller gets a hard error even though a valid screenshot exists (#1160).
-/// Returns the newest on-disk screenshot only when it was taken at or after
-/// the request started — i.e. it is THIS request landing late, not a stale
-/// image from an earlier capture. `taken_at` is second-precision, so the
-/// start is floored to the second to keep a same-second capture eligible.
-///
-/// Pure on `(started_at, latest)` for unit testing.
-fn resolve_late_screenshot(
-    started_at: chrono::DateTime<chrono::Utc>,
-    latest: Option<ScreenshotInfo>,
-) -> Option<ScreenshotInfo> {
-    let info = latest?;
-    let taken = chrono::NaiveDateTime::parse_from_str(&info.taken_at, "%Y-%m-%dT%H:%M:%SZ")
-        .ok()?
-        .and_utc();
-    let start_floor =
-        chrono::DateTime::from_timestamp(started_at.timestamp(), 0).unwrap_or(started_at);
-    if taken >= start_floor {
-        Some(info)
-    } else {
-        None
-    }
-}
-
-/// True when a captured OS window belongs to the Parish/Rundale desktop app.
-///
-/// Matched on the owning application name: the dev binary reports
-/// `parish-tauri`, a packaged build reports its bundle name (`Rundale`).
-/// Pure for unit testing the window-selection rule.
-#[cfg(any(target_os = "macos", target_os = "windows"))]
-fn is_app_window(app_name: &str) -> bool {
-    let a = app_name.to_lowercase();
-    a.contains("parish") || a.contains("rundale")
-}
-
-/// Captures the Parish/Rundale desktop window as PNG bytes using native OS
-/// screen capture (`xcap`).
-///
-/// This is the primary capture path because the gameplay MapLibre minimap is a
-/// WebGL canvas that `html-to-image` cannot read — it draws cross-origin tiles,
-/// which taint the canvas and block `toDataURL()`, so the map serialises blank
-/// (#1160 follow-up). Native capture reads the real composited pixels, map
-/// included. Picks the largest non-minimised app window. Blocking — call from
-/// `spawn_blocking`.
-///
-/// macOS/Windows only — the Linux xcap backend needs PipeWire system libs, so
-/// Linux desktop uses the html-to-image fallback instead.
-#[cfg(any(target_os = "macos", target_os = "windows"))]
-fn capture_app_window_png() -> Result<Vec<u8>, String> {
-    let windows = xcap::Window::all().map_err(|e| format!("enumerate windows: {e}"))?;
-    let mut best: Option<(u64, xcap::Window)> = None;
-    for w in windows {
-        if w.is_minimized().unwrap_or(false) {
-            continue;
-        }
-        if !is_app_window(&w.app_name().unwrap_or_default()) {
-            continue;
-        }
-        let area = u64::from(w.width().unwrap_or(0)) * u64::from(w.height().unwrap_or(0));
-        if best.as_ref().is_none_or(|(a, _)| area > *a) {
-            best = Some((area, w));
-        }
-    }
-    let (_, win) = best.ok_or("no Parish/Rundale window found to capture")?;
-    let img = win
-        .capture_image()
-        .map_err(|e| format!("capture window: {e}"))?;
-    let mut buf = Vec::new();
-    img.write_to(
-        &mut std::io::Cursor::new(&mut buf),
-        xcap::image::ImageFormat::Png,
-    )
-    .map_err(|e| format!("encode png: {e}"))?;
-    if buf.is_empty() {
-        return Err("captured image encoded to 0 bytes".into());
-    }
-    Ok(buf)
-}
-
-/// Native-capture path: grab the window pixels, write them under
-/// `<saves_dir>/screenshots/`, and update `latest_screenshot_path` so it
-/// behaves identically to the frontend round-trip from the caller's view.
-#[cfg(any(target_os = "macos", target_os = "windows"))]
-async fn do_take_screenshot_native(state: &Arc<AppState>) -> Result<ScreenshotInfo, String> {
-    let png = tokio::task::spawn_blocking(capture_app_window_png)
-        .await
-        .map_err(|e| format!("capture task join: {e}"))??;
-    let now = chrono::Utc::now();
-    let info = write_screenshot_to_disk(&state.saves_dir, &png, now)?;
-    *state.latest_screenshot_path.lock().await = Some(std::path::PathBuf::from(&info.path));
-    Ok(info)
-}
-
-/// Agent-triggered screenshot capture.
-///
-/// Tries native window capture first ([`do_take_screenshot_native`]) so the
-/// WebGL minimap appears in the image. If native capture is unavailable (no
-/// window, missing screen-recording permission, headless), it falls back to the
-/// frontend `html-to-image` round-trip in [`do_take_screenshot_frontend`]
-/// (which carries the #1160 deadline + late-capture handling). The map may be
-/// blank in the fallback, but a screenshot is still produced.
-pub(crate) async fn do_take_screenshot(
-    state: &Arc<AppState>,
-    app: &tauri::AppHandle,
-) -> Result<ScreenshotInfo, String> {
-    #[cfg(any(target_os = "macos", target_os = "windows"))]
-    match do_take_screenshot_native(state).await {
-        Ok(info) => return Ok(info),
-        Err(e) => {
-            tracing::warn!(
-                "native screenshot failed ({e}); falling back to frontend html-to-image capture"
-            );
-        }
-    }
-    do_take_screenshot_frontend(state, app).await
-}
-
-/// Frontend `html-to-image` round-trip — the fallback capture path.
-///
-/// Generates a UUID request ID, stashes a `oneshot::Sender` in
-/// `AppState::pending_screenshots`, emits `request-screenshot` to the live
-/// frontend window, and awaits the result for up to [`screenshot_timeout`]
-/// (default 45 s, `PARISH_SCREENSHOT_TIMEOUT_SECS`).
-///
-/// On emit failure the pending entry is cleaned up immediately so the map
-/// does not grow unbounded. On timeout the entry is removed; if a capture
-/// from this request landed on disk just after the deadline we return it via
-/// [`resolve_late_screenshot`] rather than erroring (#1160).
-/// The frontend delivers the result via `notify_screenshot_captured` (success)
-/// or `notify_screenshot_error` (failure).
-async fn do_take_screenshot_frontend(
-    state: &Arc<AppState>,
-    app: &tauri::AppHandle,
-) -> Result<ScreenshotInfo, String> {
-    use tokio::sync::oneshot;
-    let started_at = chrono::Utc::now();
-    let timeout = screenshot_timeout();
-    let request_id = uuid::Uuid::new_v4().to_string();
-    let (tx, rx) = oneshot::channel();
-    {
-        let mut pending = state.pending_screenshots.lock().await;
-        pending.insert(request_id.clone(), tx);
-    }
-    if let Err(e) = app.emit(
-        crate::events::EVENT_REQUEST_SCREENSHOT,
-        serde_json::json!({"request_id": request_id}),
-    ) {
-        let mut pending = state.pending_screenshots.lock().await;
-        pending.remove(&request_id);
-        return Err(e.to_string());
-    }
-    match tokio::time::timeout(timeout, rx).await {
-        Ok(Ok(result)) => result,
-        Ok(Err(_)) => Err("screenshot request cancelled".into()),
-        Err(_) => {
-            {
-                let mut pending = state.pending_screenshots.lock().await;
-                pending.remove(&request_id);
-            }
-            // The capture may have completed just after the deadline; if a
-            // fresh PNG landed on disk, return it instead of a hard error.
-            let latest = do_get_latest_screenshot(state).await.ok().flatten();
-            match resolve_late_screenshot(started_at, latest) {
-                Some(info) => Ok(info),
-                None => Err(format!(
-                    "screenshot capture timed out after {} s",
-                    timeout.as_secs()
-                )),
-            }
-        }
-    }
-}
-
-/// Called by the frontend after successfully capturing a screenshot in
-/// response to a `request-screenshot` Tauri event. Delivers the
-/// `ScreenshotInfo` through the pending oneshot channel so the waiting
-/// bridge handler can return it to the MCP client.
-///
-/// Internal round-trip callback — no HTTP equivalent on `parish-server`.
-#[tauri::command]
-pub async fn notify_screenshot_captured(
-    request_id: String,
-    info: ScreenshotInfo,
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<(), String> {
-    let mut pending = state.pending_screenshots.lock().await;
-    if let Some(tx) = pending.remove(&request_id) {
-        let _ = tx.send(Ok(info));
-    }
-    Ok(())
-}
-
-/// Called by the frontend when screenshot capture fails (e.g. `html-to-image`
-/// error or `save_screenshot` IPC error). Immediately unblocks the waiting
-/// bridge handler with the error message instead of letting it time out.
-///
-/// Internal round-trip callback — no HTTP equivalent on `parish-server`.
-#[tauri::command]
-pub async fn notify_screenshot_error(
-    request_id: String,
-    error: String,
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<(), String> {
-    let mut pending = state.pending_screenshots.lock().await;
-    if let Some(tx) = pending.remove(&request_id) {
-        let _ = tx.send(Err(error));
-    }
-    Ok(())
-}
-
-/// Formats branch list as text for the /branches command.
-pub async fn do_list_branches_text(state: &Arc<AppState>) -> Result<String, String> {
-    let db_path = {
-        let guard = state.save_path.lock().await;
-        guard
-            .as_ref()
-            .ok_or("No active save file. Use /save first.")?
-            .clone()
-    };
-    let current_id = *state.current_branch_id.lock().await;
-
-    let branches = tokio::task::spawn_blocking(move || -> Result<Vec<_>, String> {
-        let db = Database::open(&db_path).map_err(|e| e.to_string())?;
-        db.list_branches().map_err(|e| e.to_string())
-    })
-    .await
-    .map_err(|e| e.to_string())??;
-
-    let mut lines = vec!["Branches:".to_string()];
-    for b in &branches {
-        let marker = if Some(b.id) == current_id { " *" } else { "" };
-        let parent = b
-            .parent_branch_id
-            .and_then(|pid| branches.iter().find(|bb| bb.id == pid))
-            .map(|bb| format!(" (from {})", bb.name))
-            .unwrap_or_default();
-        lines.push(format!("  {}{}{}", b.name, parent, marker));
-    }
-    Ok(lines.join("\n"))
-}
-
-/// Formats branch log as text for the /log command.
-pub async fn do_branch_log_text(state: &Arc<AppState>) -> Result<String, String> {
-    let db_path = {
-        let guard = state.save_path.lock().await;
-        guard
-            .as_ref()
-            .ok_or("No active save file. Use /save first.")?
-            .clone()
-    };
-    let bid = state
-        .current_branch_id
-        .lock()
-        .await
-        .ok_or("No active branch.")?;
-
-    let log = tokio::task::spawn_blocking(move || -> Result<Vec<_>, String> {
-        let db = Database::open(&db_path).map_err(|e| e.to_string())?;
-        db.branch_log(bid).map_err(|e| e.to_string())
-    })
-    .await
-    .map_err(|e| e.to_string())??;
-
-    if log.is_empty() {
-        return Ok("No snapshots yet on this branch.".to_string());
-    }
-
-    let branch_name = state.current_branch_name.lock().await;
-    let name = branch_name.as_deref().unwrap_or("unknown");
-
-    let mut lines = vec![format!("Save log for branch '{}':", name)];
-    for (i, info) in log.iter().enumerate() {
-        let time = parish_core::persistence::format_timestamp(&info.real_time);
-        lines.push(format!("  {}. {} (game: {})", i + 1, time, info.game_time));
-    }
-    Ok(lines.join("\n"))
-}
-
-// ── Reaction commands ──────────────────────────────────────────────────────
-
-// Snippet injection validation is shared via parish_core::game_loop::is_snippet_injection_char
-// (#687 security parity). Delegating here guarantees server and Tauri use identical logic.
-pub use parish_core::game_loop::is_snippet_injection_char;
-
-/// Player reacts to an NPC message with an emoji.
-#[tauri::command]
-pub async fn react_to_message(
-    npc_name: String,
-    message_snippet: String,
-    emoji: String,
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<(), String> {
-    // Validate emoji is in the palette
-    if reactions::reaction_description(&emoji).is_none() {
-        return Err("Unknown reaction emoji.".to_string());
-    }
-
-    // Reject snippets that could inject content into NPC system prompts (#687).
-    if message_snippet.chars().any(is_snippet_injection_char) {
-        return Err("Message snippet contains disallowed characters.".to_string());
-    }
-
-    let mut npc_manager = state.npc_manager.lock().await;
-    if let Some(npc) = npc_manager.find_by_name_mut(&npc_name) {
-        let now = chrono::Utc::now();
-        npc.reaction_log.add(&emoji, &message_snippet, now);
-    }
-
-    Ok(())
-}
-
-/// Delegates to [`parish_core::game_loop::emit_npc_reactions`] (#696 slice 5).
-///
-/// `location` must be the player's location **at the time the message was
-/// sent**, captured before any `handle_game_input` call that might move the
-/// player. This prevents a race where the player moves between spawn and
-/// execution, causing reactions to be attributed to NPCs at the wrong location.
-///
-/// Pre-captures the NPC list, resolves the reaction client and feature flags,
-/// constructs a `TauriEmitter`, and delegates to the shared implementation.
-fn emit_npc_reactions(
-    player_msg_id: &str,
-    player_input: &str,
-    location: LocationId,
-    state: &Arc<AppState>,
-    app: &tauri::AppHandle,
-) {
-    let state_clone = Arc::clone(state);
-    let player_msg_id = player_msg_id.to_string();
-    let player_input = player_input.to_string();
-    let emitter: std::sync::Arc<dyn parish_core::ipc::EventEmitter> =
-        std::sync::Arc::new(crate::events::TauriEmitter::new(app.clone()));
-
-    // Persist callback: closes over Arc<AppState> and locks npc_manager to
-    // record each reaction in the NPC's reaction_log (#403).
-    let state_for_persist = Arc::clone(state);
-    let persist: parish_core::game_loop::PersistReactionFn = std::sync::Arc::new(
-        move |npc_name: String, emoji: String, player_input: String| {
-            let state = Arc::clone(&state_for_persist);
-            tokio::spawn(async move {
-                let mut npc_manager = state.npc_manager.lock().await;
-                if let Some(npc_mut) = npc_manager.find_by_name_mut(&npc_name) {
-                    npc_mut.reaction_log.add_player_message_reaction(
-                        &emoji,
-                        &player_input,
-                        chrono::Utc::now(),
-                    );
-                }
-                // Feed the per-session diversity sensor (#995).
-                npc_manager.record_reaction_emoji(&emoji);
-            });
-        },
-    );
-
-    tokio::spawn(async move {
-        // Pre-capture the NPC list at the given location (the player may have
-        // moved by the time the background task runs).
-        let (npcs_here, reaction_client, reaction_model, llm_enabled) = {
-            let npc_manager = state_clone.npc_manager.lock().await;
-            let config = state_clone.config.lock().await;
-            let base_client = state_clone.client.lock().await;
-            let npcs = npc_manager
-                .npcs_at(location)
-                .iter()
-                .map(|npc| (*npc).clone())
-                .collect::<Vec<_>>();
-            let (client, model) =
-                config.resolve_category_client(InferenceCategory::Reaction, base_client.as_ref());
-            let enabled = !config.flags.is_disabled("npc-llm-reactions");
-            (npcs, client, model, enabled)
-        };
-
-        parish_core::game_loop::emit_npc_reactions(
-            player_msg_id,
-            player_input,
-            npcs_here,
-            reaction_client,
-            reaction_model,
-            llm_enabled,
-            emitter,
-            persist,
-        );
-    });
-}
-
-// ── Demo / auto-player commands ──────────────────────────────────────────────
-
-// Demo payload types and the prompt-builder live in `parish-core::ipc::demo`
-// so the builder can be constrained to GUI-facing inputs only (issue #998).
-pub use parish_core::ipc::demo::{
-    DemoAdjacentLocation, DemoContextSnapshot, DemoNpcInfo, build_demo_context,
-};
-
-/// Demo configuration returned by `get_demo_config`.
-#[derive(serde::Serialize, Clone)]
-pub struct DemoConfigPayload {
-    pub auto_start: bool,
-    pub extra_prompt: Option<String>,
-    pub turn_pause_secs: f32,
-    pub max_turns: Option<u32>,
-}
-
-/// Returns the demo configuration (CLI flags parsed at startup).
-#[tauri::command]
-pub async fn get_demo_config(
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<DemoConfigPayload, String> {
-    let dc = &state.demo_config;
-    Ok(DemoConfigPayload {
-        auto_start: dc.auto_start,
-        extra_prompt: dc.extra_prompt.clone(),
-        turn_pause_secs: dc.turn_pause_secs,
-        max_turns: dc.max_turns,
-    })
-}
-
-/// Builds a context snapshot for the LLM demo player.
-///
-/// Returns location, time, weather, NPCs present, and adjacent locations.
-/// The `recent_log` field is empty; the frontend fills it from the text log
-/// store before passing the snapshot to `get_llm_player_action`.
-#[tauri::command]
-pub async fn get_demo_context(
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<DemoContextSnapshot, String> {
-    {
-        let config = state.config.lock().await;
-        if !config.flags.is_enabled("demo-mode") {
-            return Err("Demo mode is not active.".to_string());
-        }
-    }
-
-    // Lock order: world → npc_manager (matches AppState contract).
-    let world = state.world.lock().await;
-    let npc_manager = state.npc_manager.lock().await;
-
-    // Issue #998: build the demo snapshot from the same GUI-facing IPC
-    // payloads the frontend already consumes. Anything the GUI hides
-    // (occupation pre-intro, fog-of-war neighbours) is automatically
-    // hidden from the LLM prompt.
-    let world_snapshot = parish_core::ipc::handlers::snapshot_from_world(&world);
-    let npcs = parish_core::ipc::handlers::build_npcs_here(&world, &npc_manager);
-    let map =
-        parish_core::ipc::handlers::build_map_data(&world, state.transport.default_mode(), false);
-
-    use chrono::{Datelike, Timelike};
-    let now = world.clock.now();
-    // TODO #5: include HH:MM alongside the time-of-day word so the
-    // demo auto-player can see clock progression between turns and
-    // pick an appropriate greeting register.
-    let game_time = format!(
-        "{}, {} {} {}, {:02}:{:02} ({})",
-        now.format("%A"),
-        now.day(),
-        now.format("%B"),
-        now.year(),
-        now.hour(),
-        now.minute(),
-        world.clock.time_of_day(),
-    );
-    let season = format!("{}", world.clock.season());
-    let extra_prompt = state.demo_config.extra_prompt.clone();
-
-    Ok(build_demo_context(
-        &world_snapshot,
-        &npcs,
-        &map,
-        game_time,
-        season,
-        extra_prompt,
-    ))
-}
-
-/// Detects if a string is command-form intent description (bare verb-first
-/// pattern like "ask about X", "tell Name Y", "whisper to X") rather than
-/// first-person speech or direct command. This guard prevents the demo
-/// auto-player from leaking the LLM's internal intent reasoning as player chat.
-///
-/// Command-form intent leaks are specifically dialogue-related intent
-/// descriptions that the LLM might output during reasoning:
-/// - "ask about X", "ask the Y", "ask a Z"
-/// - "tell Name something"
-/// - "whisper to Name"
-///
-/// Movement/exploration commands like "go to X" and "look" are valid player inputs
-/// and NOT considered intent leaks.
-///
-/// Returns `true` if the text matches a dialogue intent leak pattern.
-fn is_command_form_intent_leak(text: &str) -> bool {
-    let trimmed = text.trim();
-
-    // Bare direct commands that are valid (single word, no object) are NOT intent leaks.
-    // These are legitimate player inputs for the game engine.
-    let valid_direct_commands = ["look", "wait", "go", "listen", "think"];
-    if valid_direct_commands.contains(&trimmed) {
-        return false;
-    }
-
-    // Check for dialogue-form intent leaks: bare verbs used for dialogue.
-    // These are patterns where the LLM outputs its internal reasoning about
-    // what dialogue to attempt, rather than the dialogue itself.
-    let dialogue_intent_patterns = [
-        // Dialogue-related intent leaks
-        "ask about ",
-        "ask the ",
-        "ask a ",
-        "ask if ",
-        "ask ", // bare "ask" followed by something (but not bare "ask" alone)
-        "tell ",
-        "whisper ",
-        "whisper to ",
-    ];
-
-    let lower = trimmed.to_lowercase();
-    for pattern in &dialogue_intent_patterns {
-        if lower.starts_with(pattern) {
-            return true;
-        }
-    }
-
-    false
-}
-
-/// Extracts the player action from an LLM response.
-///
-/// Handles four patterns:
-/// 1. Completion: model received `{"action": "` and completed it — response is
-///    something like `go to the mill"}`. Extract up to the closing quote.
-/// 2. Full JSON: model output `{"action": "go to the mill"}` — scan for `{`
-///    and JSON-parse from there.
-/// 3. Envelope-leak: model emitted only the JSON suffix — e.g.
-///    `action": "hello"}` or `hello"}` — strip the wrapper bits.
-/// 4. Fallback: no JSON at all — strip thinking preamble, take last line.
-fn extract_action_from_response(text: &str) -> String {
-    // Strip thinking blocks first so all patterns operate on clean text.
-    let stripped = strip_thinking_block(text);
-    let trimmed = stripped.trim();
-
-    // Pattern 1: fill-in-the-blank completion — response starts with the
-    // action text and ends with `"}` or just `"`. The model completed
-    // `{"action": "` → `go to the mill"}`.
-    // We also handle `go to the mill"` (no closing brace).
-    let completion = trimmed
-        .trim_end_matches('}')
-        .trim_end()
-        .trim_end_matches('"')
-        .trim();
-    // Valid completion: no opening brace in the extracted text (it's pure action).
-    if !completion.is_empty() && !completion.starts_with('{') && !completion.contains("action") {
-        // Check that the raw response looked like a completion (no full JSON object).
-        if !trimmed.contains("{\"action\"") && !trimmed.contains("{ \"action\"") {
-            // Guard against command-form intent leaks (e.g., "ask about ...").
-            if !is_command_form_intent_leak(completion) {
-                return completion.to_string();
-            }
-        }
-    }
-
-    // Pattern 2: full JSON object anywhere in the response.
-    let mut search = trimmed;
-    while let Some(start) = search.find('{') {
-        let candidate = &search[start..];
-        if let Ok(val) = serde_json::from_str::<serde_json::Value>(candidate)
-            && let Some(action) = val.get("action").and_then(|v| v.as_str())
-        {
-            let action = action.trim();
-            if !action.is_empty() {
-                // Guard against command-form intent leaks.
-                if !is_command_form_intent_leak(action) {
-                    return action.to_string();
-                }
-            }
-        }
-        search = &search[start + 1..];
-    }
-
-    // Pattern 3: envelope-leak — model emitted only the JSON suffix (no
-    // matching `{...}` object) such as:
-    //   action": "Good morning..."}
-    //   "action": "Good morning..."}
-    //   Good morning..."}
-    // Strip a leading `[{][\s]*["]?action["]\s*:\s*"` prefix and a trailing
-    // `"\s*}` (or bare `"}`) suffix, then return the inner text.
-    if let Some(cleaned) = strip_envelope_leak(trimmed)
-        && !cleaned.is_empty()
-    {
-        // Guard against command-form intent leaks.
-        if !is_command_form_intent_leak(&cleaned) {
-            return cleaned;
-        }
-    }
-
-    // Pattern 4: fallback — take last meaningful line from already-stripped text.
-    // Skip if the text looks like JSON (Pattern 2 already tried to extract from it).
-    if trimmed.contains("{\"") || trimmed.contains("{ \"") || trimmed.starts_with('{') {
-        // All JSON patterns either returned a leak or found nothing.
-        return String::new();
-    }
-    let fallback = trimmed.trim_matches('"').trim_matches('\'').to_string();
-    // Guard against command-form intent leaks before returning fallback.
-    if !is_command_form_intent_leak(&fallback) {
-        fallback
-    } else {
-        // If all patterns extracted a command-form intent leak, return empty
-        // to indicate the LLM failed to produce valid player input.
-        String::new()
-    }
-}
-
-/// Strips a leaked JSON envelope from `text`. Returns `Some(inner)` when at
-/// least one envelope marker (a leading `action":` prefix or a trailing `"}`
-/// suffix) was found and removed. Returns `None` if neither side looks like
-/// a leak, so the caller can apply its own fallback.
-fn strip_envelope_leak(text: &str) -> Option<String> {
-    let mut s = text.trim();
-    let mut changed = false;
-
-    // Leading wrapper: optional `{`, optional whitespace, optional `"`,
-    // literal `action`, optional `"`, whitespace, `:`, whitespace, `"`.
-    // We hand-roll this instead of pulling a regex dep.
-    let original = s;
-    let mut rest = s;
-    rest = rest.trim_start();
-    if let Some(stripped) = rest.strip_prefix('{') {
-        rest = stripped.trim_start();
-    }
-    if let Some(stripped) = rest.strip_prefix('"') {
-        rest = stripped;
-    }
-    if let Some(stripped) = rest.strip_prefix("action") {
-        let mut after = stripped;
-        if let Some(stripped) = after.strip_prefix('"') {
-            after = stripped;
-        }
-        after = after.trim_start();
-        if let Some(stripped) = after.strip_prefix(':') {
-            let mut after = stripped.trim_start();
-            if let Some(stripped) = after.strip_prefix('"') {
-                after = stripped;
-                s = after;
-                changed = true;
-            }
-        }
-    }
-    if !changed {
-        s = original;
-    }
-
-    // Trailing wrapper: optional `}`, whitespace, `"` (in reverse order
-    // since we work from the end).
-    let original_tail = s;
-    let mut tail = s.trim_end();
-    if let Some(stripped) = tail.strip_suffix('}') {
-        tail = stripped.trim_end();
-    }
-    if let Some(stripped) = tail.strip_suffix('"') {
-        tail = stripped;
-        s = tail;
-        changed = true;
-    } else {
-        s = original_tail;
-    }
-
-    if changed {
-        Some(s.trim().to_string())
-    } else {
-        None
-    }
-}
-
-/// Strips reasoning preamble from LLM responses so only the action remains.
-///
-/// Handles two patterns:
-/// 1. Tagged blocks: `<thinking>...</thinking>` / `<think>...</think>` from
-///    reasoning models (deepseek-r1, qwq). Takes everything after the last
-///    closing tag.
-/// 2. Plain-text multi-paragraph reasoning: if the response has blank-line-
-///    separated paragraphs, takes the last paragraph. This covers models that
-///    output reasoning prose before the final action without tags.
-///
-/// Falls back to the full trimmed text if neither pattern applies.
-fn strip_thinking_block(text: &str) -> &str {
-    let trimmed = text.trim();
-
-    // Strip tagged thinking blocks first.
-    for close_tag in &["</thinking>", "</think>"] {
-        if let Some(pos) = trimmed.rfind(close_tag) {
-            let after = trimmed[pos + close_tag.len()..].trim();
-            if !after.is_empty() {
-                return after;
-            }
-        }
-    }
-
-    // If the response has multiple blank-line-separated paragraphs, take the
-    // last one — reasoning models often output rationale before the action.
-    if let Some(last_para) = trimmed.rsplit("\n\n").find(|p| !p.trim().is_empty()) {
-        let candidate = last_para.trim();
-        // Only use the last paragraph if it looks like a short action (≤ 3
-        // lines), not if the whole thing is one paragraph of prose.
-        let line_count = candidate.lines().count();
-        if line_count <= 3 && candidate.len() < trimmed.len() {
-            return candidate;
-        }
-    }
-
-    // Last-line fallback: models sometimes separate reasoning from action with
-    // a single newline. If the last non-empty line is much shorter than the
-    // full response, treat it as the action.
-    let lines: Vec<&str> = trimmed.lines().filter(|l| !l.trim().is_empty()).collect();
-    if lines.len() > 1
-        && let Some(&last) = lines.last()
-    {
-        let last = last.trim();
-        if last.len() <= 200 && last.len() < trimmed.len() {
-            return last;
-        }
-    }
-
-    trimmed
-}
-
-/// Truncate `s` to at most `max_chars` characters, suffixing `...` when
-/// truncation occurs. Used to keep tracing previews bounded for the
-/// `raw_preview` field on the empty-action retry path (TODO #18).
-fn truncate_for_log(s: &str, max_chars: usize) -> String {
-    if s.chars().count() <= max_chars {
-        return s.to_string();
-    }
-    let mut out: String = s.chars().take(max_chars).collect();
-    out.push_str("...");
-    out
-}
-
-/// Builds the demo-turn system prompt for the LLM-as-player.
-///
-/// Extracted from `get_llm_player_action` so the role anchor (TODO #51)
-/// is unit-testable without driving the full Tauri command flow. The
-/// optional `extra_prompt` is appended verbatim after the "Explore
-/// naturally" paragraph — usually loaded from
-/// `mods/rundale/demo-prompt.txt`.
-fn build_demo_system_prompt(extra_prompt: Option<&str>) -> String {
-    let extra_section = extra_prompt
-        .map(|p| format!("\n\n{}", p))
-        .unwrap_or_default();
-
-    format!(
-        "You are playing Rundale, an Irish living-world simulation set in 1820. You are a \
-wandering stranger named Aiden Carney exploring the townlands of east Roscommon. The world \
-is populated by historical Irish villagers — farmers, priests, weavers, matchmakers — each \
-living their own life.\n\
-\n\
-ROLE: You are ALWAYS Aiden Carney. Speak ONLY in Aiden's voice — never as a priest, miller, \
-shopkeeper, schoolmaster, or any other local NPC. If the previous turn in the prompt ends \
-with your own line and no NPC reply, that means the NPC's reply is still in flight; you \
-still speak as Aiden on the next turn — do NOT take the NPC's side of the exchange. Do not \
-answer your own questions on the NPC's behalf, and do not roleplay an answer from a \
-villager.\n\
-\n\
-Date: 1820. Catholic Emancipation: 1829 (not yet). Famine: 1845 (not yet).\n\
-\n\
-Speak as a 1820 traveller would: plain, short, period-appropriate. Avoid modern words \
-like: fascinating, amazing, definitely, totally, decided to visit, taking in the sights, \
-healing properties.\n\
-\n\
-Explore naturally: talk to people, learn their stories, travel between locations, and \
-respond to whatever you encounter. Act as a curious outsider would.{extra}\n\
-\n\
-Respond with a JSON object containing a single field \"action\" — the text the player \
-would type into the game. Do NOT use meta-commands like \"talk to X\"; write the actual \
-words or command directly.\n\
-\n\
-NO NARRATION (TODO #47): The engine has no narrative parser. Do NOT describe what \
-you are doing in past tense, third person, or participial style. Inputs like \
-\"Walking up to the cabin, I knock gently on the door\", \"Sittin' here, I \
-notice a book half-open on the table\", or \"I'll take a seat on the bench\" \
-vanish into the dialogue path and produce no game-state change. Legal action \
-shapes are: (a) spoken dialogue in first-person present tense (\"Good \
-mornin'. Have ye news from the road?\"), (b) movement commands (\"go to The \
-Mill\"), or (c) a bare command verb the engine recognises (\"look\"). If you \
-want to do something physical, say what you would say aloud — never narrate \
-the action.\n\
-\n\
-Do NOT repeat yourself: if your last action appears in the \"Your last actions\" or \
-\"Recent events\" block of the user prompt, pick a different action — try a different \
-greeting, ask a different question, or travel somewhere new. The location description \
-is already shown to you in the prompt; you do not need to issue a bare \"look\" command.\n\
-\n\
-Do NOT mirror NPC catchphrases. The \"Recent events\" block carries NPC replies in \
-their own voice — they may use stock tags like \"Just askin', mind ye\", \"so it is\", \
-\"sure\", or \"mayhap\" as their personal vocal habits. Aiden has his own voice: use \
-plain Hiberno-English without adopting another character's verbal tics. If an NPC \
-ends every line with \"so it is\", do not start ending yours with it too.\n\
-\n\
-MOVEMENT CADENCE (TODO #1/#30): A traveller does not loiter. After 3–5 turns \
-at one location, move to a new place — pick a name from the \"You can go to: \
-...\" line in the user prompt and emit a movement command on its own (no \
-spoken line wrapped around it). Bare \"go to X\" / \"walk to X\" / \"head to X\" \
-is the correct shape: the engine parses these as movement, not dialogue. If \
-you have visited only one location in the last 5 turns, your next action \
-should be a movement command.\n\
-\n\
-WHEN ALONE (TODO #12): If the user prompt's status block contains the line \
-\"NPCs here: none\", there is nobody to hear you. Do NOT speak, ask questions, \
-roleplay knocking on doors, or wait around. Your ONLY useful action at an \
-empty location is to move. Pick a destination from the \"You can go to: ...\" \
-line and emit a bare movement command (\"go to X\" / \"walk to X\" / \"head \
-to X\"). Speaking at an empty location burns a turn and accomplishes nothing.\n\
-\n\
-Examples:\n\
-  {{\"action\": \"Good mornin' to ye. A fair day for the road.\"}}\n\
-  {{\"action\": \"I've come from up the road. What news do ye have hereabouts?\"}}\n\
-  {{\"action\": \"Might I ask about the harvest, then?\"}}\n\
-  {{\"action\": \"go to The Mill\"}}\n\
-  {{\"action\": \"walk to St. Brigid's Church\"}}\n\
-  {{\"action\": \"head to Connolly's Shop\"}}\n\
-\n\
-Your entire response must be a single JSON object — nothing before or after it.",
-        extra = extra_section,
-    )
-}
-
-/// Asks the LLM to choose the next player action given the current game context.
-///
-/// The frontend fills `ctx.recent_log` from the text log store before calling
-/// this command. Returns the trimmed action string.
-#[tauri::command]
-pub async fn get_llm_player_action(
-    ctx: DemoContextSnapshot,
-    state: tauri::State<'_, Arc<AppState>>,
-) -> Result<String, String> {
-    {
-        let config = state.config.lock().await;
-        if !config.flags.is_enabled("demo-mode") {
-            return Err("Demo mode is not active.".to_string());
-        }
-    }
-
-    // Resolve client and model (base client; no per-category override for demo).
-    let (client_opt, model) = {
-        let config = state.config.lock().await;
-        let client_guard = state.client.lock().await;
-        let model = config.model_name.clone();
-        let client = client_guard.as_ref().cloned();
-        (client, model)
-    };
-
-    let Some(client) = client_opt else {
-        return Err("No LLM client configured.".to_string());
-    };
-
-    let system_prompt = build_demo_system_prompt(ctx.extra_prompt.as_deref());
-
-    // Issue #998: render via the shared `parish_core::ipc::demo` helper so the
-    // prompt format stays in lockstep with the typed snapshot (no
-    // `Name (Title)` parens that the LLM can misread as a vocative).
-    let user_prompt = parish_core::ipc::demo::render_user_prompt(&ctx);
-
-    // Surface the constructed prompt so demo-mode logs prove what the LLM
-    // actually saw — required by the issue-998 verification flow.
-    tracing::info!(
-        location = %ctx.location_name,
-        user_prompt = %user_prompt,
-        "demo turn: prompt built"
-    );
-
-    let raw = client
-        .generate(
-            &model,
-            &user_prompt,
-            Some(&system_prompt),
-            Some(200),
-            Some(0.9),
-            None,
-        )
-        .await
-        .map_err(|e| e.to_string())?;
-
-    // Primary: extract the "action" field from JSON output.
-    // The system prompt asks for {"action": "..."}, which is robust against
-    // any amount of preamble or reasoning text the model emits before it.
-    let mut action_text = extract_action_from_response(&raw);
-    tracing::info!(
-        location = %ctx.location_name,
-        raw_len = raw.len(),
-        action = %action_text,
-        "demo turn: LLM chose action"
-    );
-
-    // TODO #18 — bounded single retry on empty action. Cycle 3 of the
-    // demo audit logged two consecutive turns where the LLM returned
-    // 137/139 chars but the parser surfaced an empty action; the
-    // player input was recorded as nothing and no NPC turn fired.
-    // Common cause: model emitted {"action": ""} or completion lacking
-    // the `action` key. Retry once at temperature 1.0 with the same
-    // prompt — most retries succeed on the bump. Bounded to one extra
-    // call so a wedged model can't pin the slot.
-    if action_text.is_empty() && !raw.trim().is_empty() {
-        tracing::warn!(
-            location = %ctx.location_name,
-            raw_len = raw.len(),
-            raw_preview = %truncate_for_log(&raw, 200),
-            "demo turn: parsed action empty despite non-empty completion; retrying once"
-        );
-        let retry_raw = client
-            .generate(
-                &model,
-                &user_prompt,
-                Some(&system_prompt),
-                Some(200),
-                Some(1.0),
-                None,
-            )
-            .await
-            .map_err(|e| e.to_string())?;
-        let retry_action = extract_action_from_response(&retry_raw);
-        if !retry_action.is_empty() {
-            tracing::info!(
-                location = %ctx.location_name,
-                raw_len = retry_raw.len(),
-                action = %retry_action,
-                "demo turn: retry produced non-empty action"
-            );
-            action_text = retry_action;
-        } else {
-            tracing::warn!(
-                location = %ctx.location_name,
-                raw_len = retry_raw.len(),
-                raw_preview = %truncate_for_log(&retry_raw, 200),
-                "demo turn: retry also produced empty action; skipping turn"
-            );
-        }
-    }
-
-    // Quality sensors — emit WARN on any structural issue in the parsed
-    // player action. These don't gate execution; they surface bugs in the
-    // demo log so the judging pass can pick them up.
-    for issue in parish_core::npc::quality::detect_all_text_issues(&action_text) {
-        tracing::warn!(
-            site = "demo-player-action",
-            kind = issue.kind.as_str(),
-            detail = %issue.detail,
-            "quality issue in LLM player action"
-        );
-    }
-
-    Ok(action_text)
-}
-
-#[cfg(test)]
-mod demo_tests {
-    use super::{
-        build_demo_system_prompt, extract_action_from_response, strip_thinking_block,
-        truncate_for_log,
-    };
-
-    /// TODO #18 — pin the failure shapes where the parser returns
-    /// empty so the retry path's gate (`action_text.is_empty()`) is
-    /// well-defined.
-    #[test]
-    fn extract_action_returns_empty_on_action_field_set_to_empty_string() {
-        // {"action": ""} — model emitted the envelope but with an
-        // empty action. Parser returns "" → retry fires.
-        assert_eq!(extract_action_from_response(r#"{"action": ""}"#), "");
-    }
-
-    #[test]
-    fn extract_action_returns_empty_on_bare_empty_input() {
-        // Bare empty string — nothing to recover. Retry would still
-        // fire on a non-empty raw completion in the actual loop.
-        assert_eq!(extract_action_from_response(""), "");
-    }
-
-    #[test]
-    fn truncate_for_log_short_string_passes_through() {
-        assert_eq!(truncate_for_log("hello", 200), "hello");
-    }
-
-    #[test]
-    fn truncate_for_log_long_string_is_clipped_with_ellipsis() {
-        let long: String = "x".repeat(500);
-        let truncated = truncate_for_log(&long, 200);
-        assert_eq!(truncated.chars().filter(|&c| c == 'x').count(), 200);
-        assert!(truncated.ends_with("..."));
-    }
-
-    #[test]
-    fn demo_system_prompt_names_aiden_carney() {
-        // TODO #51 — AC1: system prompt must explicitly name the
-        // auto-player so the model has a role anchor stronger than
-        // the generic "wandering stranger" phrasing.
-        let prompt = build_demo_system_prompt(None);
-        assert!(
-            prompt.contains("Aiden Carney"),
-            "system prompt missing player name anchor:\n{prompt}"
-        );
-    }
-
-    #[test]
-    fn demo_system_prompt_forbids_speaking_as_npc() {
-        // TODO #51 — AC2: prompt must direct the model to speak only
-        // in Aiden's voice and not roleplay an NPC's reply, even when
-        // the prior turn lacks an NPC line.
-        let prompt = build_demo_system_prompt(None);
-        assert!(
-            prompt.contains("Speak ONLY in Aiden's voice"),
-            "system prompt missing speak-only-as-Aiden directive:\n{prompt}"
-        );
-        assert!(
-            prompt.contains("never as a priest, miller, shopkeeper"),
-            "system prompt missing never-as-NPC list:\n{prompt}"
-        );
-        assert!(
-            prompt.contains("NPC's reply is still in flight"),
-            "system prompt missing in-flight-reply guidance:\n{prompt}"
-        );
-        assert!(
-            prompt.contains("do NOT take the NPC's side"),
-            "system prompt missing don't-flip-roles directive:\n{prompt}"
-        );
-    }
-
-    #[test]
-    fn demo_system_prompt_forbids_mirroring_npc_catchphrases() {
-        // TODO #26: the auto-player was adopting NPC stock tags
-        // ("Just askin', mind ye") from the recent-events buffer.
-        // Prompt must explicitly tell the model not to mirror NPC
-        // verbal tics.
-        let prompt = build_demo_system_prompt(None);
-        assert!(
-            prompt.contains("Do NOT mirror NPC catchphrases"),
-            "system prompt missing catchphrase-mirror guard:\n{prompt}"
-        );
-        assert!(
-            prompt.contains("Aiden has his own voice"),
-            "system prompt missing own-voice anchor:\n{prompt}"
-        );
-        assert!(
-            prompt.contains("Just askin', mind ye"),
-            "system prompt should name the canonical example phrase:\n{prompt}"
-        );
-    }
-
-    #[test]
-    fn demo_system_prompt_carries_movement_cadence_directive() {
-        // TODO #1/#30: auto-player produced exactly 1 movement in 38+
-        // turns because the prompt has no explicit cadence rule and
-        // movement is 1 of 4 few-shot examples. Prompt must (a) name
-        // movement as a first-class action, (b) carry a "move after
-        // N turns" cadence rule, (c) show ≥ 2 movement few-shots
-        // alongside the dialogue ones, and (d) cite all three canonical
-        // movement verbs so the model picks across them.
-        let prompt = build_demo_system_prompt(None);
-        assert!(
-            prompt.contains("MOVEMENT CADENCE"),
-            "system prompt missing movement-cadence header:\n{prompt}"
-        );
-        assert!(
-            prompt.contains("After 3–5 turns"),
-            "system prompt missing 3-5 turn cadence rule:\n{prompt}"
-        );
-        assert!(
-            prompt.contains("go to The Mill"),
-            "system prompt missing 'go to' movement example:\n{prompt}"
-        );
-        assert!(
-            prompt.contains("walk to St. Brigid's Church"),
-            "system prompt missing 'walk to' movement example:\n{prompt}"
-        );
-        assert!(
-            prompt.contains("head to Connolly's Shop"),
-            "system prompt missing 'head to' movement example:\n{prompt}"
-        );
-        assert!(
-            prompt.contains("the engine parses these as movement"),
-            "system prompt must distinguish movement commands from \
-             dialogue so the model emits bare 'go to X' rather than \
-             wrapping it in a spoken line:\n{prompt}"
-        );
-    }
-
-    #[test]
-    fn demo_system_prompt_forbids_narrative_action_style() {
-        // TODO #47: cycle 9 caught 8/18 turns in narrative form
-        // ("Walking up to the cabin, I knock gently on the door";
-        // "Sittin' here, I notice a book half-open on the table";
-        // "I'll take a seat on the bench") — all silently dropped by
-        // the engine. The prompt must (a) name the failure mode, (b)
-        // cite at least one concrete negative example so the model
-        // pattern-matches, and (c) enumerate the legal action shapes.
-        let prompt = build_demo_system_prompt(None);
-        assert!(
-            prompt.contains("NO NARRATION"),
-            "system prompt missing no-narration header:\n{prompt}"
-        );
-        assert!(
-            prompt.contains("Walking up to the cabin"),
-            "system prompt missing participial-narration negative example:\n{prompt}"
-        );
-        assert!(
-            prompt.contains("vanish into the dialogue path"),
-            "system prompt must spell out the engine's failure mode so the \
-             model has a reason to obey:\n{prompt}"
-        );
-        assert!(
-            prompt.contains("Legal action shapes are"),
-            "system prompt must enumerate the legal action shapes:\n{prompt}"
-        );
-    }
-
-    #[test]
-    fn demo_system_prompt_layers_extra_prompt() {
-        // TODO #51 — AC3: operator extra prompt must still appear.
-        let prompt = build_demo_system_prompt(Some("RUNDALE-SPECIFIC: stay east of the river."));
-        assert!(
-            prompt.contains("RUNDALE-SPECIFIC: stay east of the river."),
-            "extra prompt missing from layered system prompt:\n{prompt}"
-        );
-        // And the anchor still appears alongside the extra content.
-        assert!(prompt.contains("Speak ONLY in Aiden's voice"));
-    }
-
-    #[test]
-    fn extracts_action_from_json() {
-        let input = r#"{"action": "Good morning! How are you today?"}"#;
-        assert_eq!(
-            extract_action_from_response(input),
-            "Good morning! How are you today?"
-        );
-    }
-
-    #[test]
-    fn extracts_action_from_json_after_preamble() {
-        let input = "Let me think... However, we are a wandering stranger.\n{\"action\": \"go to the crossroads\"}";
-        assert_eq!(extract_action_from_response(input), "go to the crossroads");
-    }
-
-    #[test]
-    fn extracts_action_from_json_after_thinking_tags() {
-        let input = "<think>reasoning here</think>\n{\"action\": \"look\"}";
-        assert_eq!(extract_action_from_response(input), "look");
-    }
-
-    #[test]
-    fn falls_back_to_stripping_when_no_json() {
-        let input = "Some reasoning.\nMight I ask about the harvest, then?";
-        assert_eq!(
-            extract_action_from_response(input),
-            "Might I ask about the harvest, then?"
-        );
-    }
-
-    #[test]
-    fn strips_envelope_leak_with_action_prefix() {
-        // Live demo bug: model emits only the JSON suffix, no opening `{`.
-        let input = r#"action": "Good morning, Peig Hannigan. My name is [Your Name]. I'm just wandering."}"#;
-        assert_eq!(
-            extract_action_from_response(input),
-            "Good morning, Peig Hannigan. My name is [Your Name]. I'm just wandering."
-        );
-    }
-
-    #[test]
-    fn strips_trailing_envelope_suffix() {
-        // Live demo bug: model leaves only the trailing `"}` after a clean reply.
-        let input = r#"I'm just curious about the community."}"#;
-        assert_eq!(
-            extract_action_from_response(input),
-            "I'm just curious about the community."
-        );
-    }
-
-    #[test]
-    fn strips_thinking_block_before_action() {
-        let input =
-            "<thinking>\nI should greet the farmer.\n</thinking>\nHello there, good morning!";
-        assert_eq!(strip_thinking_block(input), "Hello there, good morning!");
-    }
-
-    #[test]
-    fn strips_think_tag_variant() {
-        let input = "<think>reasoning</think>\ngo to the mill";
-        assert_eq!(strip_thinking_block(input), "go to the mill");
-    }
-
-    #[test]
-    fn no_thinking_block_returns_trimmed() {
-        let input = "  ask Brigid about the harvest  ";
-        assert_eq!(strip_thinking_block(input), "ask Brigid about the harvest");
-    }
-
-    #[test]
-    fn only_thinking_block_falls_back_to_full() {
-        let input = "<thinking>just thinking, nothing after</thinking>";
-        assert_eq!(
-            strip_thinking_block(input),
-            "<thinking>just thinking, nothing after</thinking>"
-        );
-    }
-
-    #[test]
-    fn uses_last_closing_tag_for_nested() {
-        let input = "<thinking>outer <think>inner</think> more</thinking>\nlook around";
-        assert_eq!(strip_thinking_block(input), "look around");
-    }
-
-    #[test]
-    fn strips_plain_text_reasoning_before_action() {
-        let input = "Looking at the context, I see Peig is here. I should greet her warmly.\n\nHello Peig, good morning!";
-        assert_eq!(strip_thinking_block(input), "Hello Peig, good morning!");
-    }
-
-    #[test]
-    fn single_paragraph_returned_as_is() {
-        let input = "Hello Seamus, how goes the harvest?";
-        assert_eq!(strip_thinking_block(input), input);
-    }
-
-    #[test]
-    fn strips_single_newline_reasoning_before_action() {
-        let input = "I need to explore. The crossroads is nearby.\ngo to the crossroads";
-        assert_eq!(strip_thinking_block(input), "go to the crossroads");
-    }
-
-    #[test]
-    fn strips_multi_sentence_reasoning_single_newline() {
-        let input = "Based on my previous interaction with Peig, I should explore. The mill is unvisited.\nask about the mill";
-        assert_eq!(strip_thinking_block(input), "ask about the mill");
-    }
-
-    #[test]
-    fn is_command_form_intent_leak_rejects_ask_patterns() {
-        assert!(super::is_command_form_intent_leak(
-            "ask about the places nearby that are worth visiting"
-        ));
-        assert!(super::is_command_form_intent_leak("ask about the harvest"));
-        assert!(super::is_command_form_intent_leak("ask the priest"));
-        assert!(super::is_command_form_intent_leak("ask a stranger"));
-        assert!(super::is_command_form_intent_leak("ask if anyone knows"));
-    }
-
-    #[test]
-    fn is_command_form_intent_leak_rejects_tell_patterns() {
-        assert!(super::is_command_form_intent_leak("tell Brigid my name"));
-        assert!(super::is_command_form_intent_leak(
-            "tell the stranger something"
-        ));
-    }
-
-    #[test]
-    fn is_command_form_intent_leak_rejects_whisper_patterns() {
-        assert!(super::is_command_form_intent_leak("whisper a secret"));
-        assert!(super::is_command_form_intent_leak("whisper to Brigid"));
-    }
-
-    #[test]
-    fn is_command_form_intent_leak_accepts_look_patterns() {
-        // "look" variants are movement/exploration commands, not dialogue intent leaks.
-        assert!(!super::is_command_form_intent_leak("look at the stranger"));
-        assert!(!super::is_command_form_intent_leak("look for water"));
-    }
-
-    #[test]
-    fn is_command_form_intent_leak_accepts_movement_commands() {
-        // Movement commands are valid player inputs, NOT intent leaks.
-        assert!(!super::is_command_form_intent_leak("go to the mill"));
-        assert!(!super::is_command_form_intent_leak("go back home"));
-        assert!(!super::is_command_form_intent_leak("go into the house"));
-        assert!(!super::is_command_form_intent_leak(
-            "go towards the village"
-        ));
-        assert!(!super::is_command_form_intent_leak("walk to the mill"));
-        assert!(!super::is_command_form_intent_leak("travel to Dublin"));
-        assert!(!super::is_command_form_intent_leak("move to the window"));
-        assert!(!super::is_command_form_intent_leak("climb to the hill"));
-    }
-
-    #[test]
-    fn is_command_form_intent_leak_accepts_bare_commands() {
-        // Bare valid commands should NOT be flagged as intent leaks.
-        assert!(!super::is_command_form_intent_leak("look"));
-        assert!(!super::is_command_form_intent_leak("wait"));
-        assert!(!super::is_command_form_intent_leak("go"));
-        assert!(!super::is_command_form_intent_leak("listen"));
-        assert!(!super::is_command_form_intent_leak("think"));
-    }
-
-    #[test]
-    fn is_command_form_intent_leak_accepts_natural_speech() {
-        // Natural first-person speech should pass through.
-        assert!(!super::is_command_form_intent_leak(
-            "Good mornin'. Might I look about the village a while?"
-        ));
-        assert!(!super::is_command_form_intent_leak(
-            "I've come from up the road. What news do ye have hereabouts?"
-        ));
-        assert!(!super::is_command_form_intent_leak(
-            "Might I ask about the harvest, then?"
-        ));
-        assert!(!super::is_command_form_intent_leak(
-            "Hello there, good morning!"
-        ));
-    }
-
-    #[test]
-    fn extract_action_rejects_intent_leak_from_json() {
-        let input = r#"{"action": "ask about the places nearby that are worth visiting"}"#;
-        // Should return empty string since the intent leak is detected.
-        assert_eq!(extract_action_from_response(input), "");
-    }
-
-    #[test]
-    fn extract_action_rejects_dialogue_intent_leak_from_completion() {
-        let input = r#"ask about the places nearby that are worth visiting"}"#;
-        // Completion pattern that is a dialogue intent leak should return empty.
-        assert_eq!(extract_action_from_response(input), "");
-    }
-
-    #[test]
-    fn extract_action_accepts_natural_speech_in_json() {
-        let input = r#"{"action": "Might I ask about the harvest, then?"}"#;
-        assert_eq!(
-            extract_action_from_response(input),
-            "Might I ask about the harvest, then?"
-        );
-    }
-
-    #[test]
-    fn extract_action_accepts_bare_commands() {
-        let input = r#"{"action": "look"}"#;
-        assert_eq!(extract_action_from_response(input), "look");
-
-        let input2 = r#"{"action": "go"}"#;
-        assert_eq!(extract_action_from_response(input2), "go");
-    }
-}
-
-#[cfg(test)]
-mod cmd_tests {
-    use super::*;
-    use std::sync::Arc;
-
-    /// Builds a minimal [`AppState`] for unit tests — matches the structure
-    /// used in `parish-server` tests (`routes::tests::test_app_state`).
-    fn test_app_state() -> Arc<AppState> {
-        use crate::{
-            AppState, ConversationRuntimeState, DEBUG_EVENT_CAPACITY, DemoConfig, GameConfig,
-            UiConfigSnapshot,
-        };
-        use parish_core::inference::new_inference_log;
-        use parish_core::npc::manager::NpcManager;
-        use parish_core::world::transport::TransportConfig;
-        use parish_core::world::{DEFAULT_START_LOCATION, WorldState};
-        use tokio::sync::Mutex;
-        use tokio_util::sync::CancellationToken;
-
-        let data_dir =
-            std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("../../../mods/rundale");
-        let world =
-            WorldState::from_parish_file(&data_dir.join("world.json"), DEFAULT_START_LOCATION)
-                .unwrap();
-        let npc_manager = NpcManager::new();
-        let transport = TransportConfig::default();
-        let ui_config = UiConfigSnapshot {
-            hints_label: "test".to_string(),
-            default_accent: "#000".to_string(),
-            splash_text: String::new(),
-            active_tile_source: String::new(),
-            tile_sources: Vec::new(),
-            auto_pause_timeout_seconds: 300,
-            app_icon_url: None,
-            favicon_url: None,
-            map_overlay: None,
-            base_mod_required: false,
-        };
-        let theme_palette = parish_core::game_mod::default_theme_palette();
-        let pronunciations = Vec::new();
-        let reaction_templates = parish_core::npc::reactions::ReactionTemplates::default();
-        let game_config = GameConfig {
-            provider_name: String::new(),
-            base_url: String::new(),
-            api_key: None,
-            model_name: String::new(),
-            cloud_provider_name: None,
-            cloud_model_name: None,
-            cloud_api_key: None,
-            cloud_base_url: None,
-            improv_enabled: false,
-            max_follow_up_turns: 2,
-            idle_banter_after_secs: 25,
-            auto_pause_after_secs: 60,
-            category_provider: Default::default(),
-            category_model: Default::default(),
-            category_api_key: Default::default(),
-            category_base_url: Default::default(),
-            flags: parish_core::config::FeatureFlags::default(),
-            category_rate_limit: Default::default(),
-            active_tile_source: String::new(),
-            tile_sources: Vec::new(),
-            reveal_unexplored_locations: false,
-            auto_setup_model: None,
-        };
-        let saves_dir = std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("../../../saves");
-        let shutdown_token = CancellationToken::new();
-        let session_store: std::sync::Arc<dyn parish_core::session_store::SessionStore> =
-            std::sync::Arc::new(parish_core::session_store::DbSessionStore::new(
-                saves_dir.clone(),
-            ));
-
-        Arc::new(AppState {
-            world: Mutex::new(world),
-            npc_manager: Mutex::new(npc_manager),
-            inference_queue: Mutex::new(None),
-            client: Mutex::new(None),
-            cloud_client: Mutex::new(None),
-            conversation: Mutex::new(ConversationRuntimeState::new()),
-            debug_events: Mutex::new(std::collections::VecDeque::with_capacity(
-                DEBUG_EVENT_CAPACITY,
-            )),
-            game_events: Mutex::new(std::collections::VecDeque::with_capacity(
-                DEBUG_EVENT_CAPACITY,
-            )),
-            game_mod: None,
-            inference_log: new_inference_log(),
-            ui_config,
-            theme_palette,
-            theme_keyframes: Vec::new(),
-            static_raw_palette: None,
-            inference_failure_messages: Vec::new(),
-            idle_messages: Vec::new(),
-            pronunciations,
-            reaction_templates,
-            save_path: Mutex::new(None),
-            current_branch_id: Mutex::new(None),
-            current_branch_name: Mutex::new(None),
-            transport,
-            data_dir: data_dir.clone(),
-            saves_dir,
-            latest_screenshot_path: Mutex::new(None),
-            pending_screenshots: Mutex::new(std::collections::HashMap::new()),
-            worker_handle: Mutex::new(None),
-            editor: std::sync::Mutex::new(parish_core::ipc::editor::EditorSession::default()),
-            save_lock: Mutex::new(None),
-            runtime_processes: Mutex::new(parish_core::inference::client::RuntimeProcesses::none()),
-            inference_config: parish_core::config::InferenceConfig::default(),
-            setup_status: std::sync::Mutex::new(crate::SetupStatusSnapshot::default()),
-            wizard_in_flight: std::sync::atomic::AtomicBool::new(false),
-            language_settings: parish_core::npc::LanguageSettings::english_only(),
-            config: Mutex::new(game_config),
-            demo_config: DemoConfig::default(),
-            shutdown_token,
-            sim_cancel: Mutex::new(CancellationToken::new()),
-            session_store,
-            user_config_dir: std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR")),
-            secret_store: std::sync::Arc::new(parish_core::secret_store::InMemorySecretStore::new()),
-            inference_file_log: parish_core::inference::file_log::InferenceFileLog::disabled(),
-            chat_transcript_log: parish_core::chat_transcript::ChatTranscriptLog::disabled(),
-        })
-    }
-
-    // ── validate_input_text ─────────────────────────────────────────────────
-
-    #[test]
-    fn validate_input_accepts_normal_text() {
-        assert!(validate_input_text("ask Brigid about the harvest").is_ok());
-    }
-
-    #[test]
-    fn validate_input_trims_whitespace() {
-        let result = validate_input_text("  hello  ").unwrap();
-        assert_eq!(result, "hello");
-    }
-
-    #[test]
-    fn validate_input_allows_empty_after_trim() {
-        assert!(validate_input_text("   ").is_ok());
-    }
-
-    #[test]
-    fn validate_input_rejects_over_2000_chars() {
-        let long: String = "a".repeat(2001);
-        assert!(validate_input_text(&long).is_err());
-    }
-
-    #[test]
-    fn validate_input_accepts_exactly_2000_chars() {
-        let exactly: String = "a".repeat(2000);
-        assert!(validate_input_text(&exactly).is_ok());
-    }
-
-    // ── validate_addressed_to ───────────────────────────────────────────────
-
-    #[test]
-    fn validate_addressed_to_accepts_empty_list() {
-        assert!(validate_addressed_to(&[]).is_ok());
-    }
-
-    #[test]
-    fn validate_addressed_to_accepts_up_to_10() {
-        let names: Vec<String> = (0..10).map(|i| format!("Npc{}", i)).collect();
-        assert!(validate_addressed_to(&names).is_ok());
-    }
-
-    #[test]
-    fn validate_addressed_to_rejects_11_names() {
-        let names: Vec<String> = (0..11).map(|i| format!("Npc{}", i)).collect();
-        assert!(validate_addressed_to(&names).is_err());
-    }
-
-    #[test]
-    fn validate_addressed_to_rejects_name_over_100_chars() {
-        let long_name = "a".repeat(101);
-        assert!(validate_addressed_to(&[long_name]).is_err());
-    }
-
-    #[test]
-    fn validate_addressed_to_accepts_100_char_name() {
-        let name = "a".repeat(100);
-        assert!(validate_addressed_to(&[name]).is_ok());
-    }
-
-    // ── is_snippet_injection_char ───────────────────────────────────────────
-
-    #[test]
-    fn snippet_injection_rejects_double_quote() {
-        assert!(is_snippet_injection_char('"'));
-    }
-
-    #[test]
-    fn snippet_injection_rejects_backslash() {
-        assert!(is_snippet_injection_char('\\'));
-    }
-
-    #[test]
-    fn snippet_injection_rejects_line_separator() {
-        assert!(is_snippet_injection_char('\u{2028}'));
-    }
-
-    #[test]
-    fn snippet_injection_rejects_paragraph_separator() {
-        assert!(is_snippet_injection_char('\u{2029}'));
-    }
-
-    #[test]
-    fn snippet_injection_rejects_null_byte() {
-        assert!(is_snippet_injection_char('\0'));
-    }
-
-    #[test]
-    fn snippet_injection_accepts_normal_chars() {
-        for c in "abcdefghijklmnopqrstuvwxyz ÁÉÍÓÚ,.!?'".chars() {
-            assert!(
-                !is_snippet_injection_char(c),
-                "char {:?} should be allowed",
-                c
-            );
-        }
-    }
-
-    // ── AppState save state on fresh state ─────────────────────────────────
-
-    #[tokio::test]
-    async fn save_state_initial_is_empty() {
-        let state = test_app_state();
-        let save_path = state.save_path.lock().await;
-        let branch_id = state.current_branch_id.lock().await;
-        let branch_name = state.current_branch_name.lock().await;
-
-        let filename = save_path
-            .as_ref()
-            .and_then(|p| p.file_name())
-            .map(|n| n.to_string_lossy().to_string());
-
-        assert!(filename.is_none(), "fresh state should have no save file");
-        assert!(branch_id.is_none(), "fresh state should have no branch id");
-        assert!(
-            branch_name.is_none(),
-            "fresh state should have no branch name"
-        );
-    }
-
-    // ── conversation state ──────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn set_conversation_running_toggles_flag() {
-        let state = test_app_state();
-
-        // Initially not running
-        {
-            let conv = state.conversation.lock().await;
-            assert!(!conv.conversation_in_progress);
-        }
-
-        set_conversation_running(&state, true).await;
-
-        {
-            let conv = state.conversation.lock().await;
-            assert!(conv.conversation_in_progress);
-        }
-
-        set_conversation_running(&state, false).await;
-
-        {
-            let conv = state.conversation.lock().await;
-            assert!(!conv.conversation_in_progress);
-        }
-    }
-
-    // ── tick_inactivity does nothing when paused ────────────────────────────
-
-    #[tokio::test]
-    async fn world_clock_paused_state_has_expected_invariants() {
-        let state = test_app_state();
-
-        // Pause the world clock
-        {
-            let mut world = state.world.lock().await;
-            world.clock.pause();
-        }
-
-        // tick_inactivity needs an AppHandle which we can't construct in unit
-        // tests.  The early-return path (world is paused) never touches app,
-        // so we exercise the banter-after-silence guard indirectly via
-        // conversation state: conversation_in_progress=false, clock paused →
-        // the guard returns immediately without calling run_idle_banter.
-        // We verify world state is unchanged.
-        let (paused_before, loc_before) = {
-            let world = state.world.lock().await;
-            (world.clock.is_paused(), world.player_location)
-        };
-        assert!(paused_before, "clock should be paused before tick");
-
-        // We can't call tick_inactivity here because it needs tauri::AppHandle.
-        // Instead, confirm the state invariants hold so future tests that mock
-        // AppHandle can call tick_inactivity against this base.
-        let paused_after = {
-            let world = state.world.lock().await;
-            world.clock.is_paused()
-        };
-        assert!(paused_after, "paused flag should still be set");
-        assert_eq!(
-            state.world.lock().await.player_location,
-            loc_before,
-            "location should be unchanged"
-        );
-    }
-
-    // ── world snapshot consistency ─────────────────────────────────────────
-
-    #[tokio::test]
-    async fn world_state_loads_kilteevan_as_start_location() {
-        let state = test_app_state();
-        let world = state.world.lock().await;
-        let loc_name = world
-            .current_location_data()
-            .map(|d| d.name.as_str())
-            .unwrap_or("unknown");
-        // Default start location for Rundale is Kilteevan Village
-        assert_eq!(loc_name, "Kilteevan Village");
-    }
-
-    #[tokio::test]
-    async fn discover_save_files_returns_ok_for_missing_saves_dir() {
-        let state = test_app_state();
-        let world = state.world.lock().await;
-        let nonexistent_dir = std::path::Path::new("/tmp/rundale_test_nonexistent_saves_dir");
-        let saves = parish_core::persistence::picker::discover_saves(nonexistent_dir, &world.graph);
-        // Missing dir should return empty vec, not panic
-        assert!(
-            saves.is_empty(),
-            "discover_saves should return empty vec for missing directory"
-        );
-    }
-
-    #[tokio::test]
-    async fn get_world_snapshot_inner_returns_start_location() {
-        let state = test_app_state();
-        let world = state.world.lock().await;
-        let npc_manager = state.npc_manager.lock().await;
-        let snapshot = get_world_snapshot_inner(&world, Some(&npc_manager), &state.pronunciations);
-        assert!(
-            !snapshot.location_name.is_empty(),
-            "location name should be populated"
-        );
-        assert_eq!(snapshot.location_name, "Kilteevan Village");
-    }
-
-    // ── Screenshot helpers ────────────────────────────────────────────────
-
-    /// A 1×1 transparent PNG, as the smallest valid payload to round-trip.
-    const ONE_PIXEL_PNG_B64: &str = "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNkAAIAAAoAAv/lxKUAAAAASUVORK5CYII=";
-
-    fn one_pixel_data_url() -> String {
-        format!("data:image/png;base64,{ONE_PIXEL_PNG_B64}")
-    }
-
-    #[test]
-    fn decode_data_url_png_accepts_well_formed_url() {
-        let bytes = decode_data_url_png(&one_pixel_data_url()).unwrap();
-        // PNG magic header is 8 bytes starting with 0x89 'P' 'N' 'G'.
-        assert_eq!(&bytes[..4], &[0x89, b'P', b'N', b'G']);
-    }
-
-    #[test]
-    fn decode_data_url_png_rejects_wrong_prefix() {
-        let err = decode_data_url_png("data:image/jpeg;base64,xxxx").unwrap_err();
-        assert!(err.contains("data:image/png;base64,"), "got: {err}");
-    }
-
-    #[test]
-    fn decode_data_url_png_rejects_invalid_base64() {
-        let err = decode_data_url_png("data:image/png;base64,***not-base64***").unwrap_err();
-        assert!(err.contains("base64 decode failed"), "got: {err}");
-    }
-
-    #[test]
-    fn write_screenshot_to_disk_creates_file_and_reports_size() {
-        let tmp = tempfile::tempdir().unwrap();
-        let bytes = decode_data_url_png(&one_pixel_data_url()).unwrap();
-        let now = chrono::DateTime::parse_from_rfc3339("2026-05-09T12:34:56Z")
-            .unwrap()
-            .with_timezone(&chrono::Utc);
-        let info = write_screenshot_to_disk(tmp.path(), &bytes, now).unwrap();
-
-        let path = std::path::PathBuf::from(&info.path);
-        assert!(path.exists(), "PNG should be written to {}", info.path);
-        assert!(
-            info.path.ends_with("parish-2026-05-09T12-34-56Z.png"),
-            "filename should embed the timestamp; got {}",
-            info.path
-        );
-        assert_eq!(info.size_bytes, bytes.len() as u64);
-        assert_eq!(info.taken_at, "2026-05-09T12:34:56Z");
-
-        // The directory was auto-created.
-        assert!(tmp.path().join("screenshots").is_dir());
-    }
-
-    #[tokio::test]
-    async fn do_save_screenshot_round_trips_through_app_state() {
-        // Override saves_dir to a fresh tempdir so the test is hermetic
-        // (otherwise the screenshot would land under the workspace `saves/`
-        // dir and pollute later runs).
-        let tmp = tempfile::tempdir().unwrap();
-        let mut state = test_app_state();
-        let s = std::sync::Arc::get_mut(&mut state)
-            .expect("test_app_state must hand back a unique Arc");
-        s.saves_dir = tmp.path().to_path_buf();
-
-        let info = do_save_screenshot(&state, one_pixel_data_url())
-            .await
-            .expect("screenshot should save");
-
-        assert!(std::path::PathBuf::from(&info.path).exists());
-
-        // The latest_screenshot_path is now populated.
-        let latest = state.latest_screenshot_path.lock().await.clone();
-        assert_eq!(
-            latest.map(|p| p.to_string_lossy().to_string()),
-            Some(info.path.clone()),
-        );
-
-        // get_latest_screenshot reads the same file back.
-        let read_back = do_get_latest_screenshot(&state)
-            .await
-            .expect("read back must succeed")
-            .expect("a screenshot should be cached");
-        assert_eq!(read_back.path, info.path);
-        assert_eq!(read_back.size_bytes, info.size_bytes);
-    }
-
-    #[tokio::test]
-    async fn do_get_latest_screenshot_returns_none_when_unset() {
-        let state = test_app_state();
-        let info = do_get_latest_screenshot(&state).await.unwrap();
-        assert!(info.is_none(), "no screenshot taken yet → None");
-    }
-
-    #[tokio::test]
-    async fn do_get_latest_screenshot_returns_none_when_file_missing() {
-        let mut state = test_app_state();
-        let s = std::sync::Arc::get_mut(&mut state).unwrap();
-        // Point at a path that doesn't exist on disk.
-        *s.latest_screenshot_path.get_mut() =
-            Some(std::path::PathBuf::from("/no/such/parish-screenshot.png"));
-        let info = do_get_latest_screenshot(&state).await.unwrap();
-        assert!(info.is_none(), "missing file on disk → None");
-    }
-
-    // ── #1160 screenshot deadline + late-capture fallback ────────────────────
-
-    #[test]
-    fn screenshot_timeout_default_env_and_garbage() {
-        // One test (not several) so the shared process env var is mutated
-        // sequentially and cannot race a sibling test running in parallel.
-        // SAFETY: edition-2024 env mutation; confined to this single test.
-        unsafe { std::env::remove_var("PARISH_SCREENSHOT_TIMEOUT_SECS") };
-        assert_eq!(screenshot_timeout().as_secs(), 45, "default is 45 s");
-
-        unsafe { std::env::set_var("PARISH_SCREENSHOT_TIMEOUT_SECS", "90") };
-        assert_eq!(screenshot_timeout().as_secs(), 90, "env override honoured");
-
-        // Zero and non-numeric values fall back to the default.
-        unsafe { std::env::set_var("PARISH_SCREENSHOT_TIMEOUT_SECS", "0") };
-        assert_eq!(screenshot_timeout().as_secs(), 45, "zero ignored");
-        unsafe { std::env::set_var("PARISH_SCREENSHOT_TIMEOUT_SECS", "soon") };
-        assert_eq!(screenshot_timeout().as_secs(), 45, "garbage ignored");
-
-        unsafe { std::env::remove_var("PARISH_SCREENSHOT_TIMEOUT_SECS") };
-    }
-
-    fn info_at(taken_at: &str) -> ScreenshotInfo {
-        ScreenshotInfo {
-            path: "/tmp/parish-shot.png".into(),
-            taken_at: taken_at.into(),
-            size_bytes: 1234,
-        }
-    }
-
-    #[test]
-    fn resolve_late_screenshot_returns_capture_from_this_request() {
-        let started = chrono::DateTime::parse_from_rfc3339("2026-05-31T23:59:33Z")
-            .unwrap()
-            .with_timezone(&chrono::Utc);
-        // Same second (taken_at is floored to the second) → eligible.
-        let info = resolve_late_screenshot(started, Some(info_at("2026-05-31T23:59:33Z")));
-        assert!(
-            info.is_some(),
-            "same-second late capture should be returned"
-        );
-        // Strictly newer → eligible.
-        let info = resolve_late_screenshot(started, Some(info_at("2026-05-31T23:59:40Z")));
-        assert!(info.is_some(), "newer capture should be returned");
-    }
-
-    #[test]
-    fn resolve_late_screenshot_rejects_stale_or_missing() {
-        let started = chrono::DateTime::parse_from_rfc3339("2026-05-31T23:59:33Z")
-            .unwrap()
-            .with_timezone(&chrono::Utc);
-        // Older than the request start → stale, must not be returned.
-        assert!(resolve_late_screenshot(started, Some(info_at("2026-05-31T23:59:10Z"))).is_none());
-        // No screenshot on disk at all.
-        assert!(resolve_late_screenshot(started, None).is_none());
-        // Unparseable timestamp → treated as not eligible.
-        assert!(resolve_late_screenshot(started, Some(info_at("not-a-date"))).is_none());
-    }
-
-    #[cfg(any(target_os = "macos", target_os = "windows"))]
-    #[test]
-    fn is_app_window_matches_parish_and_rundale_only() {
-        // Dev binary and packaged bundle names (case-insensitive).
-        assert!(is_app_window("parish-tauri"));
-        assert!(is_app_window("Parish"));
-        assert!(is_app_window("Rundale"));
-        assert!(is_app_window("RUNDALE"));
-        // Unrelated apps are ignored so we never capture the wrong window.
-        assert!(!is_app_window("Safari"));
-        assert!(!is_app_window("Terminal"));
-        assert!(!is_app_window(""));
-    }
-
-    // ── #9 sim-cancel preemption ─────────────────────────────────────────────
-
-    /// Snapshotting `sim_cancel`, then calling the fire-and-replace block
-    /// from `do_submit_input`, must cancel the snapshot AND leave a fresh
-    /// (uncancelled) token in place for the next sim cycle.
-    #[tokio::test]
-    async fn sim_cancel_fires_snapshot_and_replaces_token() {
-        let state = test_app_state();
-
-        // Snapshot the current sim_cancel — this is what Tier 2/3 spawn
-        // captures before player input arrives.
-        let snapshot = state.sim_cancel.lock().await.clone();
-        assert!(!snapshot.is_cancelled(), "fresh token starts uncancelled");
-
-        // Inline the fire-and-replace block from do_submit_input. We don't
-        // call do_submit_input directly because it needs a tauri::AppHandle.
-        {
-            let mut sc = state.sim_cancel.lock().await;
-            sc.cancel();
-            *sc = tokio_util::sync::CancellationToken::new();
-        }
-
-        // The snapshot is cancelled — any in-flight Tier 2/3 call that
-        // captured it will see is_cancelled() == true and bail.
-        assert!(
-            snapshot.is_cancelled(),
-            "captured snapshot must observe cancellation"
-        );
-
-        // The new token replacing it is *not* cancelled — the next sim
-        // cycle captures this one cleanly.
-        let next = state.sim_cancel.lock().await.clone();
-        assert!(
-            !next.is_cancelled(),
-            "replacement token must be fresh / uncancelled"
-        );
-    }
-
-    // ── #1196 — do_create_branch validation gate ────────────────────────────
-
-    /// `do_create_branch` must reject names with disallowed characters before
-    /// acquiring any locks or touching the database.
-    #[tokio::test]
-    async fn create_branch_rejects_invalid_name_before_db() {
-        let state = test_app_state();
-        let result = do_create_branch(&state, "bad/name!!", 1).await;
-        let err = result.expect_err("expected validation error for 'bad/name!!'");
-        assert!(
-            err.contains("Invalid branch name"),
-            "error should mention invalid branch name, got: {err}"
-        );
-    }
-
-    /// Names with more than 64 characters must be rejected.
-    #[tokio::test]
-    async fn create_branch_rejects_too_long_name() {
-        let state = test_app_state();
-        let long_name = "a".repeat(65);
-        let result = do_create_branch(&state, &long_name, 1).await;
-        let err = result.expect_err("expected validation error for 65-char name");
-        assert!(
-            err.contains("Invalid branch name"),
-            "error should mention invalid branch name, got: {err}"
-        );
-    }
-
-    /// A valid name should pass validation; failure beyond that is acceptable
-    /// (e.g. missing save file) — but the error must NOT be a validation error.
-    #[tokio::test]
-    async fn create_branch_accepts_valid_name() {
-        let state = test_app_state();
-        let result = do_create_branch(&state, "my branch 1", 1).await;
-        // No save file in test state → expected to fail with "No active save file"
-        // but NOT with "Invalid branch name".
-        match result {
-            Ok(_) => { /* valid name was accepted end-to-end */ }
-            Err(e) => {
-                assert!(
-                    !e.contains("Invalid branch name"),
-                    "valid name must not fail validation, got: {e}"
-                );
-            }
-        }
-    }
-}
+//!
+//! This file is the Rust-2018 hub: it declares the command-family submodules
+//! and re-exports every `#[tauri::command]` and public helper so `lib.rs`,
+//! the `invoke_handler` registration, and all external `use` paths are unchanged.
+//!
+//! Submodule layout:
+//! - `snapshot`  — read-only world/map/NPC/theme/debug snapshot commands
+//! - `setup`     — BYOK onboarding and local inference setup commands
+//! - `input`     — player input submission, validation, and dispatch
+//! - `movement`  — handle_movement, handle_look
+//! - `saves`     — save/load/branch/new-game persistence commands
+//! - `admin`     — debug snapshot builder, inference rebuild, inactivity tick, bug report
+//! - `screenshot`— screenshot capture, storage, and round-trip callbacks
+//! - `reactions` — emoji reactions and background NPC reaction emission
+//! - `demo`      — demo/auto-player commands and prompt helpers
+//! - `cmd_tests` — shared test helpers (test-only, cfg(test))
+
+pub mod admin;
+pub(crate) mod cmd_tests;
+pub mod demo;
+pub mod input;
+pub(crate) mod movement;
+pub mod reactions;
+pub mod saves;
+pub mod screenshot;
+pub mod setup;
+pub mod snapshot;
+
+// ── Re-exports: snapshot ──────────────────────────────────────────────────────
+pub use snapshot::get_world_snapshot_inner;
+
+// ── Re-exports: setup ─────────────────────────────────────────────────────────
+pub use setup::{LocalSetupArgs, OnboardingOptions};
+pub(crate) use setup::{do_get_onboarding_options, do_start_local_inference_setup};
+
+// ── Re-exports: input ─────────────────────────────────────────────────────────
+pub use input::{validate_addressed_to, validate_input_text};
+pub(crate) use input::{MAX_ADDRESSED_TO, MAX_TARGETS, do_submit_input, handle_game_input};
+
+// ── Re-exports: saves ─────────────────────────────────────────────────────────
+pub use saves::{do_create_branch, do_load_branch, do_new_game};
+pub(crate) use saves::{do_branch_log_text, do_list_branches_text, do_save_game};
+
+// ── Re-exports: admin ─────────────────────────────────────────────────────────
+pub use admin::rebuild_inference_inner;
+pub(crate) use admin::{build_app_debug_snapshot, do_submit_bug_report, tick_inactivity};
+
+// ── Re-exports: screenshot ────────────────────────────────────────────────────
+pub use screenshot::{ScreenshotInfo, decode_data_url_png, write_screenshot_to_disk};
+pub(crate) use screenshot::{do_get_latest_screenshot, do_save_screenshot, do_take_screenshot};
+
+// ── Re-exports: reactions ─────────────────────────────────────────────────────
+pub use reactions::is_snippet_injection_char;
+
+// ── Re-exports: demo ──────────────────────────────────────────────────────────
+pub use demo::{DemoAdjacentLocation, DemoConfigPayload, DemoContextSnapshot, DemoNpcInfo, build_demo_context};
