+++ parish/crates/parish-tauri/src/lib.rs
@@ -1364,41 +1364,41 @@ pub fn run() {
     tauri::Builder::default()
         .manage(Arc::clone(&state))
         .invoke_handler(tauri::generate_handler![
-            commands::get_world_snapshot,
-            commands::get_map,
-            commands::get_npcs_here,
-            commands::get_theme,
-            commands::get_ui_config,
-            commands::get_debug_snapshot,
-            commands::get_setup_snapshot,
-            commands::toggle_fullscreen,
-            commands::set_provider_config,
-            commands::validate_provider_config,
-            commands::get_provider_config,
-            commands::clear_provider_config,
-            commands::list_byok_env_keys,
-            commands::list_preset_models,
-            commands::list_available_providers,
-            commands::get_onboarding_options,
-            commands::start_local_inference_setup,
-            commands::submit_input,
-            commands::discover_save_files,
-            commands::save_game,
-            commands::load_branch,
-            commands::create_branch,
-            commands::new_save_file,
-            commands::new_game,
-            commands::get_save_state,
-            commands::submit_bug_report,
-            commands::react_to_message,
-            commands::get_demo_config,
-            commands::get_demo_context,
-            commands::get_llm_player_action,
-            commands::save_screenshot,
-            commands::get_latest_screenshot,
-            commands::take_screenshot,
-            commands::notify_screenshot_captured,
-            commands::notify_screenshot_error,
+            commands::snapshot::get_world_snapshot,
+            commands::snapshot::get_map,
+            commands::snapshot::get_npcs_here,
+            commands::snapshot::get_theme,
+            commands::snapshot::get_ui_config,
+            commands::snapshot::get_debug_snapshot,
+            commands::snapshot::get_setup_snapshot,
+            commands::snapshot::toggle_fullscreen,
+            commands::setup::set_provider_config,
+            commands::setup::validate_provider_config,
+            commands::setup::get_provider_config,
+            commands::setup::clear_provider_config,
+            commands::setup::list_byok_env_keys,
+            commands::setup::list_preset_models,
+            commands::setup::list_available_providers,
+            commands::setup::get_onboarding_options,
+            commands::setup::start_local_inference_setup,
+            commands::input::submit_input,
+            commands::saves::discover_save_files,
+            commands::saves::save_game,
+            commands::saves::load_branch,
+            commands::saves::create_branch,
+            commands::saves::new_save_file,
+            commands::saves::new_game,
+            commands::saves::get_save_state,
+            commands::admin::submit_bug_report,
+            commands::reactions::react_to_message,
+            commands::demo::get_demo_config,
+            commands::demo::get_demo_context,
+            commands::demo::get_llm_player_action,
+            commands::screenshot::save_screenshot,
+            commands::screenshot::get_latest_screenshot,
+            commands::screenshot::take_screenshot,
+            commands::screenshot::notify_screenshot_captured,
+            commands::screenshot::notify_screenshot_error,
             editor_commands::editor_list_mods,
             editor_commands::editor_open_mod,
             editor_commands::editor_get_snapshot,
